==> 1.All_10mers_libraries_QC_and_DESeq2.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Set Global Variables here

```{r}
# Please change the working directory
Working_dir <- "~/Desktop/AVATAR/PADIT-seq-Manuscript/Code_Availability_Revision/"

# Please be sure to install these libraries
library(pheatmap); library(ggplot2); library(DESeq2); library(ggrepel); library("stringr")

# FDR cutoff (do not change)
Cutoff_Value <- 0.05

```


# QC of all-10mers PADIT-seq libraries for EGR1 and HOXD13

```{r}
# Read file containing read counts per TFBS
ReadCounts_perTFBS <- read.table(paste(paste(Working_dir, "Input_Files", sep = "/"), "ReadCounts_perTFBS_all10mers_2TFs_EGR1-HOXD13.txt", sep = "/"), header = TRUE)

################################################################# Correlation Heatmap
data <- ReadCounts_perTFBS

# 
CorrelationMatrix_pearson <- matrix(nrow=ncol(data),ncol=ncol(data))
CorrelationMatrix_spearman <- matrix(nrow=ncol(data),ncol=ncol(data))
for(i in 1:ncol(data)){
  for(j in 1:ncol(data)){
    CorrelationMatrix_pearson[i,j] <- cor(data[,i],data[,j],method="pearson")
    CorrelationMatrix_spearman[i,j] <- cor(data[,i],data[,j],method="spearman")}}
colnames(CorrelationMatrix_pearson)<-as.vector(colnames(data))
rownames(CorrelationMatrix_pearson)<-as.vector(colnames(data))

# 
pheatmap(CorrelationMatrix_pearson,cluster_rows=TRUE,cluster_cols=TRUE,
         show_rownames=TRUE,show_colnames=TRUE,display_numbers=TRUE,main="R (pearson)", fontsize = 18)

#################################################################  PCA
y_pca_all <- prcomp(data, center=TRUE, scale.=TRUE)
PCA_Matrix <- data.frame(y_pca_all$rotation)

# Plotting PCA
a <- ggplot(data.frame(y_pca_all$rotation), aes(x=PC1, y=PC2, label=rownames(PCA_Matrix))) + geom_point(color=c(rep("black", times=3), rep("red", times=3), rep("blue", times=3)), size=2) + ggtitle("PCA (All)") + 
  xlab(paste("PC1 [",summary(y_pca_all)$importance[2,1]*100,"% Variance Explained]")) + ylab(paste("PC2 [",summary(y_pca_all)$importance[2,2]*100,"% Variance Explained]")) +
  theme(text=element_text(size=18)) + geom_text_repel(size = 8, box.padding = unit(0.5, "lines"), point.padding = unit(0.35, "lines"), colour = "black")
print(a)

# Plotting PCA
a <- ggplot(data.frame(y_pca_all$rotation), aes(x=PC1, y=PC2, label=rownames(PCA_Matrix))) + geom_point(color=c(rep("black", times=3), rep("red", times=3), rep("blue", times=3)), size=5) + ggtitle("PCA (All)") + 
  xlab(paste("PC1 [",summary(y_pca_all)$importance[2,1]*100,"% Variance Explained]")) + ylab(paste("PC2 [",summary(y_pca_all)$importance[2,2]*100,"% Variance Explained]")) +
  theme(text=element_text(size=18)) 
print(a)

```


# DESeq2 analysis of HOXD13
```{r}
NumReplicates <- 3

################################################################# DESeq2 analysis 
data_HOXD13 <- ReadCounts_perTFBS

# Filter for minimum count
data_HOXD13 <- data_HOXD13[which(apply(data_HOXD13, 1, min) > 5),]

# Format files to run DESeq2 
condition <- relevel(as.factor(c(rep("RNA", NumReplicates), 
                                 rep("ZIF268", NumReplicates), 
                                 rep("HOXD13", NumReplicates))), ref = "RNA")
colData <- data.frame(row.names=as.character(1:(NumReplicates*3)), condition=condition)
colnames(data_HOXD13) <- rownames(colData)

# Run DESeq2 
dds <- DESeqDataSetFromMatrix(countData = data_HOXD13, colData=colData, design = ~condition)
dds_results_HOXD13 <- DESeq(dds, fitType='parametric')

# DESeq2 output
dds_results_HOXD13_df <- data.frame(results(dds_results_HOXD13, contrast = c("condition","HOXD13","RNA")))

# MPRA Activity 
dds_results_HOXD13_df$MPRAactivity <- (dds_results_HOXD13_df$padj<Cutoff_Value & dds_results_HOXD13_df$log2FoldChange>0)*1 + (dds_results_HOXD13_df$padj<Cutoff_Value & dds_results_HOXD13_df$log2FoldChange<0)*-1
table(dds_results_HOXD13_df$MPRAactivity)

# Write files
write.table(dds_results_HOXD13_df, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_HOXD13_df.txt", sep = "/"), col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

```


# DESeq2 analysis of EGR1 
```{r}
NumReplicates <- 3

################################################################# DESeq2 analysis 
data_ZIF268 <- ReadCounts_perTFBS

# Filter for minimum count
data_ZIF268 <- data_ZIF268[which(apply(data_ZIF268, 1, FUN = min) > 5),]

# Format files to run DESeq2 
condition <- relevel(as.factor(c(rep("RNA", NumReplicates), 
                                 rep("ZIF268", NumReplicates), 
                                 rep("HOXD13", NumReplicates))), ref = "RNA")
colData <- data.frame(row.names=as.character(1:(NumReplicates*3)), condition=condition)
colnames(data_ZIF268) <- rownames(colData)

# Run DESeq2 
dds <- DESeqDataSetFromMatrix(countData = data_ZIF268, colData=colData, design = ~condition)
dds_results_ZIF268 <- DESeq(dds, fitType='local')

# DESeq2 output
dds_results_ZIF268_df <- data.frame(results(dds_results_ZIF268, contrast = c("condition","ZIF268","RNA")))

# MPRA Activity 
dds_results_ZIF268_df$MPRAactivity <- (dds_results_ZIF268_df$padj<Cutoff_Value & dds_results_ZIF268_df$log2FoldChange>0)*1 + (dds_results_ZIF268_df$padj<Cutoff_Value & dds_results_ZIF268_df$log2FoldChange<0)*-1
table(dds_results_ZIF268_df$MPRAactivity)

# Write files
write.table(dds_results_ZIF268_df, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_ZIF268_df.txt", sep = "/"), col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

#################################################################
```

# QC of all-10mers PADIT-seq libraries for NKX2.5, TBX5, Pho4 and Cbf1

```{r}
# Read file containing read counts per TFBS
ReadCounts_perTFBS <- read.table(paste(paste(Working_dir, "Input_Files", sep = "/"), "ReadCounts_perTFBS_all10mers_4TFs-NKX2.5-TBX5-Pho4-Cbf1.txt", sep = "/"), header = TRUE)

################################################################# Correlation Heatmap
data <- ReadCounts_perTFBS

# 
CorrelationMatrix_pearson <- matrix(nrow=ncol(data),ncol=ncol(data))
CorrelationMatrix_spearman <- matrix(nrow=ncol(data),ncol=ncol(data))
for(i in 1:ncol(data)){
  for(j in 1:ncol(data)){
    CorrelationMatrix_pearson[i,j] <- cor(data[,i],data[,j],method="pearson")
    CorrelationMatrix_spearman[i,j] <- cor(data[,i],data[,j],method="spearman")}}
colnames(CorrelationMatrix_pearson)<-as.vector(colnames(data))
rownames(CorrelationMatrix_pearson)<-as.vector(colnames(data))

# 
pheatmap(CorrelationMatrix_pearson,cluster_rows=TRUE,cluster_cols=TRUE,
         show_rownames=TRUE,show_colnames=TRUE,display_numbers=TRUE,main="R (pearson)", fontsize = 18)

#################################################################  PCA
# PCA
y_pca_all <- prcomp(data, center=TRUE, scale.=TRUE)
PCA_Matrix <- data.frame(y_pca_all$rotation)

# Plotting PCA
a <- ggplot(data.frame(y_pca_all$rotation), aes(x=PC1, y=PC2, label=rownames(PCA_Matrix))) + geom_point(color=c(rep("black", times=3), rep("red", times=4), rep("blue", times=5), rep("green", times=5), rep("yellow", times=5), rep("orange", times=5)), size=2) + ggtitle("PCA (All)") + 
  xlab(paste("PC1 [",summary(y_pca_all)$importance[2,1]*100,"% Variance Explained]")) + ylab(paste("PC2 [",summary(y_pca_all)$importance[2,2]*100,"% Variance Explained]")) +
  theme(text=element_text(size=18)) + geom_text_repel(size = 8, box.padding = unit(0.5, "lines"), point.padding = unit(0.35, "lines"), colour = "black")
print(a)

# Plotting PCA
a <- ggplot(data.frame(y_pca_all$rotation), aes(x=PC1, y=PC2, label=rownames(PCA_Matrix))) + geom_point(color=c(rep("black", times=3), rep("red", times=4), rep("blue", times=5), rep("green", times=5), rep("yellow", times=5), rep("orange", times=5)), size=5) + ggtitle("PCA (All)") + 
  xlab(paste("PC1 [",summary(y_pca_all)$importance[2,1]*100,"% Variance Explained]")) + ylab(paste("PC2 [",summary(y_pca_all)$importance[2,2]*100,"% Variance Explained]")) +
  theme(text=element_text(size=18)) 
print(a)

```


# DESeq2 analysis of NKX2.5
```{r}
# subset relevant columns
data_NKX2.5 <- ReadCounts_perTFBS_DF[,c(paste("NoDBD_R", 1:7, sep = ""),
                                        paste("ALFA_NKX2.5_R", 1:5, sep = ""))]

# Format files to run DESeq2 
condition <- relevel(as.factor(c(rep("RNA", 7), 
                                 rep("NKX2.5", 5))), ref = "RNA")
LibraryPrep_batch <- relevel(as.factor(c(rep("Exp1", 3), rep("Exp2", 9))), ref = "Exp2")
colData <- data.frame(row.names=as.character(1:12), LibraryPrep_batch = LibraryPrep_batch, condition = condition)
colnames(data_NKX2.5) <- rownames(colData)

# Run DESeq2 
dds <- DESeqDataSetFromMatrix(countData = data_NKX2.5, colData=colData, design = ~LibraryPrep_batch+condition)
dds_results_NKX2.5 <- DESeq(dds, fitType='local')

# DESeq2 output
dds_results_NKX2.5_df <- data.frame(results(dds_results_NKX2.5, contrast = c("condition","NKX2.5","RNA")))

# MPRA Activity 
dds_results_NKX2.5_df$MPRAactivity <- (dds_results_NKX2.5_df$padj<Cutoff_Value & dds_results_NKX2.5_df$log2FoldChange>0)*1 + (dds_results_NKX2.5_df$padj<Cutoff_Value & dds_results_NKX2.5_df$log2FoldChange<0)*-1
table(dds_results_NKX2.5_df$MPRAactivity)

# Write files
write.table(dds_results_NKX2.5_df, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_NKX2.5_df.txt", sep = "/"), col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

```



# DESeq2 analysis of TBX5
```{r}
# subset relevant columns
data_TBX5 <- ReadCounts_perTFBS_DF[,c(paste("NoDBD_R", 1:7, sep = ""),
                                        paste("TBX5_ALFA_R", 1:5, sep = ""))]

# Format files to run DESeq2 
condition <- relevel(as.factor(c(rep("RNA", 7), 
                                 rep("TBX5", 5))), ref = "RNA")
LibraryPrep_batch <- relevel(as.factor(c(rep("Exp1", 3), rep("Exp2", 9))), ref = "Exp2")
colData <- data.frame(row.names=as.character(1:12), LibraryPrep_batch = LibraryPrep_batch, condition = condition)
colnames(data_TBX5) <- rownames(colData)

# Run DESeq2 
dds <- DESeqDataSetFromMatrix(countData = data_TBX5, colData=colData, design = ~LibraryPrep_batch+condition)
dds_results_TBX5 <- DESeq(dds, fitType='local')

# DESeq2 output
dds_results_TBX5_df <- data.frame(results(dds_results_TBX5, contrast = c("condition","TBX5","RNA")))

# MPRA Activity 
dds_results_TBX5_df$MPRAactivity <- (dds_results_TBX5_df$padj<Cutoff_Value & dds_results_TBX5_df$log2FoldChange>0)*1 + (dds_results_TBX5_df$padj<Cutoff_Value & dds_results_TBX5_df$log2FoldChange<0)*-1
table(dds_results_TBX5_df$MPRAactivity)

# Write files
write.table(dds_results_TBX5_df, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_TBX5_df.txt", sep = "/"), col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

```


# DESeq2 analysis of Pho4
```{r}
# subset relevant columns
data_PHO4 <- ReadCounts_perTFBS_DF[,c(paste("NoDBD_R", 1:7, sep = ""),
                                        paste("ALFA_PHO4_R", 1:5, sep = ""))]

# Format files to run DESeq2 
condition <- relevel(as.factor(c(rep("RNA", 7), 
                                 rep("PHO4", 5))), ref = "RNA")
LibraryPrep_batch <- relevel(as.factor(c(rep("Exp1", 3), rep("Exp2", 9))), ref = "Exp2")
colData <- data.frame(row.names=as.character(1:12), LibraryPrep_batch = LibraryPrep_batch, condition = condition)
colnames(data_PHO4) <- rownames(colData)

# Run DESeq2 
dds <- DESeqDataSetFromMatrix(countData = data_PHO4, colData=colData, design = ~LibraryPrep_batch+condition)
dds_results_PHO4 <- DESeq(dds, fitType='local')

# DESeq2 output
dds_results_PHO4_df <- data.frame(results(dds_results_PHO4, contrast = c("condition","PHO4","RNA")))

# MPRA Activity 
dds_results_PHO4_df$MPRAactivity <- (dds_results_PHO4_df$padj<Cutoff_Value & dds_results_PHO4_df$log2FoldChange>0)*1 + (dds_results_PHO4_df$padj<Cutoff_Value & dds_results_PHO4_df$log2FoldChange<0)*-1
table(dds_results_PHO4_df$MPRAactivity)

# Write files
write.table(dds_results_PHO4_df, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_PHO4_df.txt", sep = "/"), col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

```


# DESeq2 analysis of Cbf1
```{r}
# subset relevant columns
data_CBF1 <- ReadCounts_perTFBS_DF[,c(paste("NoDBD_R", 1:7, sep = ""),
                                        paste("ALFA_CBF1_R", 1:5, sep = ""))]

# Format files to run DESeq2 
condition <- relevel(as.factor(c(rep("RNA", 7), 
                                 rep("CBF1", 5))), ref = "RNA")
LibraryPrep_batch <- relevel(as.factor(c(rep("Exp1", 3), rep("Exp2", 9))), ref = "Exp2")
colData <- data.frame(row.names=as.character(1:12), LibraryPrep_batch = LibraryPrep_batch, condition = condition)
colnames(data_CBF1) <- rownames(colData)

# Run DESeq2 
dds <- DESeqDataSetFromMatrix(countData = data_CBF1, colData=colData, design = ~LibraryPrep_batch+condition)
dds_results_CBF1 <- DESeq(dds, fitType='local')

# DESeq2 output
dds_results_CBF1_df <- data.frame(results(dds_results_CBF1, contrast = c("condition","CBF1","RNA")))

# MPRA Activity 
dds_results_CBF1_df$MPRAactivity <- (dds_results_CBF1_df$padj<Cutoff_Value & dds_results_CBF1_df$log2FoldChange>0)*1 + (dds_results_CBF1_df$padj<Cutoff_Value & dds_results_CBF1_df$log2FoldChange<0)*-1
table(dds_results_CBF1_df$MPRAactivity)

# Write files
write.table(dds_results_CBF1_df, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_CBF1_df.txt", sep = "/"), col.names = TRUE, row.names = TRUE, quote = FALSE, sep = "\t")

```
