==> 2.Compare_EGR1_PADIT-seq_to_PBM_HT-SELEX.Rmd <==
---
title: "R Notebook"
output: html_notebook
---
# Set Global Variables here

```{r}
# Please change the working directory
Working_dir <- "~/Desktop/AVATAR/PADIT-seq-Manuscript/Code_Availability_Revision/"

# Please be sure to install these libraries
library(pheatmap); library(ggplot2); library(DESeq2); library(ggrepel); library("stringr")

# FDR cutoff (do not change)
Cutoff_Value <- 0.05
```



# Obtain EGR1 9-mer PADIT-seq log2foldchange values from 10-mer data
# Caution: This is going to take a while (~7 hours) (Sorry - I only had to do this once so I did not bother optimizing the code)
```{r}
# # Read ZIF268 PADIT-seq data
# dds_results_ZIF268_df <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_ZIF268_df.txt", sep = "/"), header = TRUE)
# dds_results_ZIF268_df$TFBS <- rownames(dds_results_ZIF268_df)
# 
# # Read and format ZIF268 PBM data
# V1 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_Egr1", sep = "/"), "Zif268_v1_kf0.5_9mers_111111111.txt", sep = "/"), header = TRUE)
# V2 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_Egr1", sep = "/"), "Zif268_v2_kf0.5_9mers_111111111.txt", sep = "/"), header = TRUE)
# TEMP <- merge(V1, V2, by = "X9.mer")
# TEMP$E.score <- apply(TEMP[, c("E.score.x", "E.score.y")], 1, max)
# TEMP$Z.score <- apply(TEMP[, c("Z.score.x", "Z.score.y")], 1, min)
# ZIF268_PBM_Escores <- TEMP[, c("X9.mer", "X9.mer.1.x", "E.score", "Z.score")]; rm(TEMP)
# colnames(ZIF268_PBM_Escores) <- c("X9.mer", "X9.mer.1", "E.score", "Z.score")
# 
# # 
# for(k in 0:1) # The 2 9-mers overlapping each 10-mer
# {
#   #
#   ZIF268_PBM_Escores_PADIT <- data.frame()
#   for(i in 1:nrow(ZIF268_PBM_Escores)) # For each register, calculate the median 9-mer scores
#   {
#     Current_8mers <- ZIF268_PBM_Escores[i,]
#   
#     Fwd <- dds_results_ZIF268_df[which(substr(dds_results_ZIF268_df$TFBS, k+1, k+9) == Current_8mers$X9.mer[1]),]
#     Fwd.tmp <- Fwd[order(Fwd$log2FoldChange, decreasing = TRUE),]
#     Fwd <- Fwd.tmp; rm(Fwd.tmp)
#     if(nrow(Fwd) > 0)
#     {
#       Fwd_subset <- Fwd[ceiling(nrow(Fwd)/2), c("log2FoldChange", "padj", "MPRAactivity")]
#     }
#     else
#     {
#       Fwd_subset <- as.data.frame(t(c(-100, 1, 0)))
#       colnames(Fwd_subset) <- c("log2FoldChange", "padj", "MPRAactivity")
#     }
#   
#     Rev <- dds_results_ZIF268_df[which(substr(dds_results_ZIF268_df$TFBS, k+1, k+9) == Current_8mers$X9.mer.1[1]),]
#     Rev.tmp <- Rev[order(Rev$log2FoldChange, decreasing = TRUE),]
#     Rev <- Rev.tmp; rm(Rev.tmp)
#     if(nrow(Rev) > 0)
#     {
#       Rev_subset <- Rev[ceiling(nrow(Rev)/2), c("log2FoldChange", "padj", "MPRAactivity")]
#     }
#     else
#     {
#       Rev_subset <- as.data.frame(t(c(-100, 1, 0)))
#       colnames(Rev_subset) <- c("log2FoldChange", "padj", "MPRAactivity")
#     }
#   
#     if((Fwd_subset$log2FoldChange >= Rev_subset$log2FoldChange) | (nrow(Rev_subset) == 0))
#     {
#       Current_8mers <- cbind(Current_8mers, Fwd_subset)
#     }
#     if((Fwd_subset$log2FoldChange < Rev_subset$log2FoldChange) | (nrow(Fwd_subset) == 0))
#     {
#       Current_8mers <- cbind(Current_8mers, Rev_subset)
#     }
#   
#     ZIF268_PBM_Escores_PADIT <- rbind(ZIF268_PBM_Escores_PADIT, Current_8mers)
#   
#     # remove all the variable in current iteration
#     variables_to_remove = c("Current_8mers", "Fwd", "Fwd_subset", "Rev", "Rev_subset")
#     for (var_name in variables_to_remove)
#     {
#       if (exists(var_name))
#       {
#         rm(list=var_name)
#       }
#     }
#   }
#   
#   # Store the files
#   if(k == 0){ write.table(ZIF268_PBM_Escores_PADIT, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "ZIF268_all9mers_1_9_median.txt", sep = "/"), col.names = TRUE, row.names = FALSE, quote = FALSE, sep = "\t") }
#   if(k == 1){ write.table(ZIF268_PBM_Escores_PADIT, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "ZIF268_all9mers_2_10_median.txt", sep = "/"), col.names = TRUE, row.names = FALSE, quote = FALSE, sep = "\t") }
# }

```


# Plot comparison to PBM E-scores
```{r}
# Read and format ZIF268 PBM data
V1 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_Egr1", sep = "/"), "Zif268_v1_kf0.5_9mers_111111111.txt", sep = "/"), header = TRUE)
V2 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_Egr1", sep = "/"), "Zif268_v2_kf0.5_9mers_111111111.txt", sep = "/"), header = TRUE)
TEMP <- merge(V1, V2, by = "X9.mer")
TEMP$E.score <- apply(TEMP[, c("E.score.x", "E.score.y")], 1, max)
TEMP$Z.score <- apply(TEMP[, c("Z.score.x", "Z.score.y")], 1, min)
ZIF268_PBM_Escores <- TEMP[, c("X9.mer", "X9.mer.1.x", "E.score", "Z.score")]; rm(TEMP)
colnames(ZIF268_PBM_Escores) <- c("X9.mer", "X9.mer.1", "E.score", "Z.score")

# Read PADIT-seq log2foldchanges
Register_1_9 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "ZIF268_all9mers_1_9_median.txt", sep = "/"), header = TRUE)
Register_2_10 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "ZIF268_all9mers_2_10_median.txt", sep = "/"), header = TRUE)

# Re-define MPRAactivity
Register_1_9$MPRAactivity <- (Register_1_9$log2FoldChange > 0 & Register_1_9$padj < Cutoff_Value)*1
Register_2_10$MPRAactivity <- (Register_2_10$log2FoldChange > 0 & Register_2_10$padj < Cutoff_Value)*1

# Number of red points?
table(Register_1_9$MPRAactivity)
table(Register_2_10$MPRAactivity)

# Fraction of red points?
sum((Register_1_9$MPRAactivity == 1)*1)/nrow(Register_1_9)
sum((Register_2_10$MPRAactivity == 1)*1)/nrow(Register_2_10)

# Format
colnames(Register_1_9) <- paste(colnames(Register_1_9), "1_9", sep = "_")
colnames(Register_2_10) <- paste(colnames(Register_2_10), "2_10", sep = "_")
TEMP1 <- merge(Register_1_9, Register_2_10, 
               by.x = "X9.mer_1_9", by.y = "X9.mer_2_10")
ZIF268_PBM_Escores_PADIT <- TEMP1; rm(TEMP1)
ZIF268_PBM_Escores_PADIT$Mean_log2FC <- apply(ZIF268_PBM_Escores_PADIT[,c("log2FoldChange_1_9", "log2FoldChange_2_10")], 1, min)
ZIF268_PBM_Escores_PADIT$padj <- apply(ZIF268_PBM_Escores_PADIT[,c("padj_1_9", "padj_2_10")], 1, max)
ZIF268_PBM_Escores_PADIT$MPRAactivity <- apply(ZIF268_PBM_Escores_PADIT[,c("MPRAactivity_1_9", "MPRAactivity_2_10")], 1, min)

# Plot (against e-scores)
TEMP <- merge(ZIF268_PBM_Escores, ZIF268_PBM_Escores_PADIT, by.x = "X9.mer", by.y = "X9.mer_1_9")
ZIF268_PBM_Escores_PADIT_subset <- TEMP
ZIF268_PBM_Escores_PADIT_subset$Colour <- "black"
ZIF268_PBM_Escores_PADIT_subset$Colour[which(ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)] <- "red"
plot(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     cex = 0.4, pch = 19, cex.axis = 2, cex.lab = 2,
     col = "black", 
     xlim = c(-0.5, 0.5), ylim = c(-0.5, 2.8), 
     xlab = "E-score", ylab = "PADIT-seq log2 (ALFA-ZIF268 / No-DBD)")
points(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "red")], 
       ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "red")], 
     cex = 0.4, pch = 19,
     col = "red")
abline(h = 0, col = "red")
abline(v = 0.45, col = "red")
table(ZIF268_PBM_Escores_PADIT_subset$Colour)

# Plot (against z-scores)
plot(ZIF268_PBM_Escores_PADIT_subset$Z.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     cex = 0.4, pch = 19, cex.axis = 2, cex.lab = 2,
     col = "black", 
     # xlim = c(-0.5, 0.5), 
     ylim = c(-0.5, 2.8), 
     xlab = "Z-score", ylab = "PADIT-seq log2 (ALFA-ZIF268 / No-DBD)")
points(ZIF268_PBM_Escores_PADIT_subset$Z.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "red")], 
       ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "red")], 
     cex = 0.4, pch = 19,
     col = "red")
abline(h = 0, col = "red")
cor.test(ZIF268_PBM_Escores_PADIT_subset$Z.score, ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC)

```




################################################## Add HT-SELEX, ProBound and FIMO data for AUROC analysis
```{r}
################################################## Add HT-SELEX data
# Read HT-SELEX enriched 9-mers
table_1 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle1_enriched_9mers.txt", sep = "/"), header = TRUE)
table_2 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle2_enriched_9mers.txt", sep = "/"), header = TRUE)
table_3 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle3_enriched_9mers.txt", sep = "/"), header = TRUE)
table_4 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle4_enriched_9mers.txt", sep = "/"), header = TRUE)

# Calculate Observed / Expected Ratio
table_1$Cycle1 <- table_1$ObservedCount/table_1$ExpectedCount
table_2$Cycle2 <- table_2$ObservedCount/table_2$ExpectedCount
table_3$Cycle3 <- table_3$ObservedCount/table_3$ExpectedCount
table_4$Cycle4 <- table_4$ObservedCount/table_4$ExpectedCount

# merge -1 
table_1 <- table_1[, c("Kmer", "Cycle1")]
table_2 <- table_2[, c("Kmer", "Cycle2")]
table_3 <- table_3[, c("Kmer", "Cycle3")]
table_4 <- table_4[, c("Kmer", "Cycle4")]
HT_SELEX <- merge(merge(merge(table_1, table_2, by = "Kmer"), table_3, by = "Kmer"), table_4, by = "Kmer")

# merge -2
TEMP1 <- merge(ZIF268_PBM_Escores_PADIT_subset, HT_SELEX, by.x = "X9.mer", by.y = "Kmer", all.x = TRUE)
TEMP2 <- merge(TEMP1, HT_SELEX, by.x = "X9.mer.1", by.y = "Kmer", all.x = TRUE)
TEMP2[is.na(TEMP2)] <- 0

# Keep orientation that is larger
TEMP2$Cycle1 <- apply(TEMP2[, c("Cycle1.x", "Cycle1.y")], 1, max)
TEMP2$Cycle2 <- apply(TEMP2[, c("Cycle2.x", "Cycle2.y")], 1, max)
TEMP2$Cycle3 <- apply(TEMP2[, c("Cycle3.x", "Cycle3.y")], 1, max)
TEMP2$Cycle4 <- apply(TEMP2[, c("Cycle4.x", "Cycle4.y")], 1, max)
ZIF268_PBM_Escores_PADIT_HTSELEX <- TEMP2[, which(!colnames(TEMP2) %in% c("Cycle1.x", "Cycle1.y", "Cycle2.x", "Cycle2.y", "Cycle3.x", "Cycle3.y", "Cycle4.x", "Cycle4.y"))]

################################################## Add ProBound data
# _NNNNN
All9mers_1 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers__NNNNN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_1) <- c("kmer", "rel_affinity1")

# N_NNNN
All9mers_2 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_N_NNNN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_2) <- c("kmer", "rel_affinity2")

# NN_NNN
All9mers_3 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NN_NNN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_3) <- c("kmer", "rel_affinity3")

# NNN_NN
All9mers_4 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NNN_NN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_4) <- c("kmer", "rel_affinity4")

# NNNN_N
All9mers_5 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NNNN_N_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_5) <- c("kmer", "rel_affinity5")

# NNNNN_
All9mers_6 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NNNNN__ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_6) <- c("kmer", "rel_affinity6")

### Merge
All9mers <- All9mers_1
All9mers$rel_affinity2 <- All9mers_2$rel_affinity2
All9mers$rel_affinity3 <- All9mers_3$rel_affinity3
All9mers$rel_affinity4 <- All9mers_4$rel_affinity4
All9mers$rel_affinity5 <- All9mers_5$rel_affinity5
All9mers$rel_affinity6 <- All9mers_6$rel_affinity6
All9mers$trimmed_kmer <- substr(All9mers$kmer, 1, 9)
All9mers$rel_affinity <- apply(All9mers[, paste("rel_affinity", 1:6, sep = "")], 1, max)
ProBound_results <- All9mers[, c("trimmed_kmer", "rel_affinity")]

# Merge
TEMP1 <- merge(ZIF268_PBM_Escores_PADIT_HTSELEX, ProBound_results, 
               by.x = "X9.mer", by.y = "trimmed_kmer")
TEMP2 <- merge(TEMP1, ProBound_results, 
               by.x = "X9.mer.1", by.y = "trimmed_kmer")
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound <- TEMP2
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound$log2_rel_affinity <- log2(apply(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound[, c("rel_affinity.x", "rel_affinity.y")], 1, max))
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound <- ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound[, which(!colnames(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound) %in% c("rel_affinity.x", "rel_affinity.y"))]

################################################## Add FIMO data
# 1-9
All9mers_1 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/FIMO", sep = "/"), "All_9mers.Match.MA0162.3.pos.1.9.txt", sep = "/"), header = FALSE)[, c(1, 8)]
colnames(All9mers_1) <- c("sequence_name", "p.value1")
# 2-10
All9mers_2 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/FIMO", sep = "/"), "All_9mers.Match.MA0162.3.pos.2.10.txt", sep = "/"), header = FALSE)[, c(1, 8)]
colnames(All9mers_2) <- c("sequence_name", "p.value2")
# 3-11
All9mers_3 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/FIMO", sep = "/"), "All_9mers.Match.MA0162.3.pos.3.11.txt", sep = "/"), header = FALSE)[, c(1, 8)]
colnames(All9mers_3) <- c("sequence_name", "p.value3")
# 4-12
All9mers_4 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/FIMO", sep = "/"), "All_9mers.Match.MA0162.3.pos.4.12.txt", sep = "/"), header = FALSE)[, c(1, 8)]
colnames(All9mers_4) <- c("sequence_name", "p.value4")
# 5-13
All9mers_5 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/FIMO", sep = "/"), "All_9mers.Match.MA0162.3.pos.5.13.txt", sep = "/"), header = FALSE)[, c(1, 8)]
colnames(All9mers_5) <- c("sequence_name", "p.value5")
# 6-14
All9mers_6 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/FIMO", sep = "/"), "All_9mers.Match.MA0162.3.pos.6.14.txt", sep = "/"), header = FALSE)[, c(1, 8)]
colnames(All9mers_6) <- c("sequence_name", "p.value6")

### Merge-1
PWM_FIMO <- merge(merge(merge(merge(merge(All9mers_1, All9mers_2, by = "sequence_name"), All9mers_3, by = "sequence_name"), All9mers_4, by = "sequence_name"), All9mers_5, by = "sequence_name"), All9mers_6, by = "sequence_name")
PWM_FIMO$FIMO.p.value <- apply(PWM_FIMO[, paste("p.value", 1:6, sep = "")], 1, min)
PWM_FIMO <- PWM_FIMO[, c("sequence_name", "FIMO.p.value")]

### Merge-2
TEMP1 <- merge(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound, PWM_FIMO, 
               by.x = "X9.mer.1", by.y = "sequence_name")
TEMP2 <- merge(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound, PWM_FIMO, 
               by.x = "X9.mer", by.y = "sequence_name")
TEMP3 <- rbind(TEMP1, TEMP2)
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO<- TEMP3
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$FIMO.p.value <- -log10(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$FIMO.p.value)

```




################################################## Plot AUROC (All background)
```{r}
library("DescTools")
Columns_To_Plot <- c("E.score", "Cycle1", "Cycle2", "Cycle3", "Cycle4", "log2_rel_affinity", "FIMO.p.value")
# Columns_To_Plot <- c("E.score", "Z.score")
for(k in 1:length(Columns_To_Plot))
{
  thresholds <- seq(min(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO[, Columns_To_Plot[k]]), 
                    max(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO[, Columns_To_Plot[k]]), length.out = 5000)
  fraction_red <- vector()
  fraction_black <- vector()
  for(i in 1:length(thresholds))
  {
    ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$AboveThreshold <- (ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO[, Columns_To_Plot[k]] > thresholds[i])*1
    fraction_red[i] <- nrow(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity == 1 & ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$AboveThreshold == 1),])
    fraction_black[i] <- nrow(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity != 1 & ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$AboveThreshold == 1),])
  }
  
  # print AUC
  DF <- cbind(xvalues = fraction_black/(nrow(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO) - sum(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity)), 
              yvalues = fraction_red/sum(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity))
  DF <- as.data.frame(DF)
  DF <- DF[order(DF$xvalues, decreasing = FALSE),]
  DF <- rbind(DF, c(1, 1))
  print(suppressWarnings(AUC(DF$xvalues, DF$yvalues)))
  
  # plot
  if(k == 1)
  {
    # Plot AUC curve
    plot(DF$xvalues, DF$yvalues, type = "l", col = "black")
    abline(a = 0, b = 1, col = "black")
  }
  if(k > 1 & k < 6)
  {
    # Plot AUC curve
    points(DF$xvalues, DF$yvalues, type = "l", col = "cyan3")
  }
  if(k == 6)
  {
    # Plot AUC curve
    points(DF$xvalues, DF$yvalues, type = "l", col = "red")
  }
  if(k == 7)
  {
    # Plot AUC curve
    points(DF$xvalues, DF$yvalues, type = "l", col = "green")
  }
}


```


# Specificity versus Sensitivity of HT-SELEX data
```{r}
# Observed versus Expected threshold
ratio_threshold <- 3

# Define the active 8-mers
ZIF268_PBM_Escores_PADIT_subset_sig9mers <- c(ZIF268_PBM_Escores_PADIT_subset$X9.mer[which(ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)], 
                                              ZIF268_PBM_Escores_PADIT_subset$X9.mer.1[which(ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)])
ZIF268_PBM_Escores_PADIT_subset_sig9mers <- unique(ZIF268_PBM_Escores_PADIT_subset_sig9mers)

# Read HT-SELEX enriched 8-mers
table_1 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle1_enriched_9mers.txt", sep = "/"), header = TRUE)
table_2 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle2_enriched_9mers.txt", sep = "/"), header = TRUE)
table_3 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle3_enriched_9mers.txt", sep = "/"), header = TRUE)
table_4 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1", sep = "/"), "EGR1_HT-SELEX_Cycle4_enriched_9mers.txt", sep = "/"), header = TRUE)

# filter
table_1 <- table_1[which(table_1$ObservedCount/table_1$ExpectedCount > ratio_threshold),]
table_2 <- table_2[which(table_2$ObservedCount/table_2$ExpectedCount > ratio_threshold),]
table_3 <- table_3[which(table_3$ObservedCount/table_3$ExpectedCount > ratio_threshold),]
table_4 <- table_4[which(table_4$ObservedCount/table_4$ExpectedCount > ratio_threshold),]

# Add relevant columns
ZIF268_PBM_Escores_PADIT_subset$Cycle1 <- (ZIF268_PBM_Escores_PADIT_subset$X9.mer %in% table_1$Kmer | ZIF268_PBM_Escores_PADIT_subset$X9.mer.1 %in% table_1$Kmer)*1 
ZIF268_PBM_Escores_PADIT_subset$Cycle2 <- (ZIF268_PBM_Escores_PADIT_subset$X9.mer %in% table_2$Kmer | ZIF268_PBM_Escores_PADIT_subset$X9.mer.1 %in% table_2$Kmer)*1 
ZIF268_PBM_Escores_PADIT_subset$Cycle3 <- (ZIF268_PBM_Escores_PADIT_subset$X9.mer %in% table_3$Kmer | ZIF268_PBM_Escores_PADIT_subset$X9.mer.1 %in% table_3$Kmer)*1 
ZIF268_PBM_Escores_PADIT_subset$Cycle4 <- (ZIF268_PBM_Escores_PADIT_subset$X9.mer %in% table_4$Kmer | ZIF268_PBM_Escores_PADIT_subset$X9.mer.1 %in% table_4$Kmer)*1 

# Number of rows
num_enriched_HTSELEX <- c(sum(ZIF268_PBM_Escores_PADIT_subset$Cycle1),
                          sum(ZIF268_PBM_Escores_PADIT_subset$Cycle2),
                          sum(ZIF268_PBM_Escores_PADIT_subset$Cycle3),
                          sum(ZIF268_PBM_Escores_PADIT_subset$Cycle4))

# Specificity (red)
vec1 <- c(nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle1 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]),
          nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle2 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]),
          nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle3 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]),
          nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle4 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]))
vec1 <- vec1/num_enriched_HTSELEX

# Sensitivity (blue)
vec2 <- c(nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle1 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]),
          nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle2 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]),
          nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle3 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]),
          nrow(ZIF268_PBM_Escores_PADIT_subset[which(ZIF268_PBM_Escores_PADIT_subset$Cycle4 == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1),]))
vec2 <- vec2/length(ZIF268_PBM_Escores_PADIT_subset_sig9mers)

# 
plot(1:4, vec1, type = "b", pch = 19, col = "red", ylim = c(0, 1), xlab = "SELEX Round", cex = 4) # Fraction of true positives
points(1:4, vec2, type = "b", pch = 19, col = "blue", cex = 4) # Fraction of false negatives
```



# Plot comparison to HT-SELEX
```{r}
# Observed versus Expected threshold
ratio_threshold <- 3

# Select Cycle
ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle <- ZIF268_PBM_Escores_PADIT_subset$Cycle2

# Plot - 1
ZIF268_PBM_Escores_PADIT_subset$Colour <- "black"
ZIF268_PBM_Escores_PADIT_subset$Colour[which(ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)] <- "red"
ZIF268_PBM_Escores_PADIT_subset$Colour[which(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle == 1)] <- "cyan3"
plot(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     cex = 0.4, pch = 19, cex.axis = 2, cex.lab = 2,
     col = "black", 
     xlim = c(-0.5, 0.5), ylim = c(-0.5, 2.8), 
     xlab = "E-score", ylab = "PADIT-seq log2 (ALFA-ZIF268 / No-DBD)")
points(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "red")], 
       ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "red")], 
     cex = 0.4, pch = 19,
     col = "red")
points(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "cyan3")],
       ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "cyan3")],
     cex = 0.4, pch = 19,
     col = "cyan3")
abline(h = 0, col = "red")
table(ZIF268_PBM_Escores_PADIT_subset$Colour)

# Plot - 2
plot(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour == "black")], 
     cex = 0.4, pch = 19, cex.axis = 2, cex.lab = 2,
     col = "black", 
     xlim = c(-0.5, 0.5), ylim = c(-0.5, 2.8), 
     xlab = "E-score", ylab = "PADIT-seq log2 (ALFA-ZIF268 / No-DBD)")
points(ZIF268_PBM_Escores_PADIT_subset$E.score[which(ZIF268_PBM_Escores_PADIT_subset$Colour != "black")],
       ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Colour != "black")],
     cex = 0.4, pch = 19,
     col = ZIF268_PBM_Escores_PADIT_subset$Colour[which(ZIF268_PBM_Escores_PADIT_subset$Colour != "black")])
abline(h = 0, col = "red")
table(ZIF268_PBM_Escores_PADIT_subset$Colour)

# Boxplot
boxplot(ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle == 0 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)],
        ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)],
        notch = TRUE, outline = FALSE, col = c("red", "blue"))
wilcox.test(ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle == 0 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)],
            ZIF268_PBM_Escores_PADIT_subset$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle == 1 & ZIF268_PBM_Escores_PADIT_subset$MPRAactivity == 1)])

# Table
ZIF268_PBM_Escores_PADIT_subset$MPRAactivity[is.na(ZIF268_PBM_Escores_PADIT_subset$MPRAactivity)] <- 0
table(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle, ZIF268_PBM_Escores_PADIT_subset$MPRAactivity)

##########################
ZIF268_PBM_Escores_PADIT_subset$InteractionTerm <- interaction(ZIF268_PBM_Escores_PADIT_subset$Selected_Cycle, ZIF268_PBM_Escores_PADIT_subset$MPRAactivity)
boxplot(ZIF268_PBM_Escores_PADIT_subset$E.score ~ ZIF268_PBM_Escores_PADIT_subset$InteractionTerm,
        notch = TRUE, outline = FALSE, col = "yellow")
pairwise.wilcox.test(ZIF268_PBM_Escores_PADIT_subset$E.score, ZIF268_PBM_Escores_PADIT_subset$InteractionTerm)
table(ZIF268_PBM_Escores_PADIT_subset$InteractionTerm)

```





# Compare PADIT-seq and MITOMI-derived Kd
```{r}
# Read MITOMI KD data
MITOMI_KD <- read.table(paste(paste(Working_dir, "Input_Files", sep = "/"), "ZIF268_MITOMI_Kd_kmer9.txt", sep = "/"), header = TRUE)

# Merge
TEMP1 <- merge(MITOMI_KD, ZIF268_PBM_Escores_PADIT, by.x = "kmer9", by.y = "X9.mer_1_9")[, c("kmer9", "MITOMI_Kd", "Mean_log2FC", "MPRAactivity", "E.score_1_9", "Z.score_1_9")]
TEMP2 <- merge(MITOMI_KD, ZIF268_PBM_Escores_PADIT, by.x = "kmer9", by.y = "X9.mer.1_1_9")[, c("kmer9", "MITOMI_Kd", "Mean_log2FC", "MPRAactivity", "E.score_1_9", "Z.score_1_9")]
dds_results_ZIF268_MITOMI_df <- rbind(TEMP1, TEMP2)

############################################################ plot (against PADIT-seq)
dds_results_ZIF268_MITOMI_df$Colour <- "black"
dds_results_ZIF268_MITOMI_df$Colour[which(dds_results_ZIF268_MITOMI_df$MPRAactivity == 1)] <- "red"

# plot with best fit line
plot(-log10(dds_results_ZIF268_MITOMI_df$MITOMI_Kd), 
     dds_results_ZIF268_MITOMI_df$Mean_log2FC, 
     pch = 19, cex = 2, col = dds_results_ZIF268_MITOMI_df$Colour,
     xlab = "-log10(MITOMI KD)", ylab = "PADIT-seq log2FC")

# Correlation
cor.test(-log10(dds_results_ZIF268_MITOMI_df$MITOMI_Kd), 
         dds_results_ZIF268_MITOMI_df$Mean_log2FC, method = "pearson")
cor.test(-log10(dds_results_ZIF268_MITOMI_df$MITOMI_Kd), 
         dds_results_ZIF268_MITOMI_df$Mean_log2FC, method = "spearman")

############################################################ plot (against PBM)
dds_results_ZIF268_MITOMI_df$Colour <- "black"
dds_results_ZIF268_MITOMI_df$Colour[which(dds_results_ZIF268_MITOMI_df$MPRAactivity == 1)] <- "red"

# plot with best fit line
plot(-log10(dds_results_ZIF268_MITOMI_df$MITOMI_Kd), 
     dds_results_ZIF268_MITOMI_df$Z.score_1_9, 
     pch = 19, cex = 2, col = dds_results_ZIF268_MITOMI_df$Colour,
     xlab = "-log10(MITOMI KD)", ylab = "PADIT-seq log2FC")

# Correlation
cor.test(-log10(dds_results_ZIF268_MITOMI_df$MITOMI_Kd), 
         dds_results_ZIF268_MITOMI_df$Z.score_1_9, method = "pearson")
cor.test(-log10(dds_results_ZIF268_MITOMI_df$MITOMI_Kd), 
         dds_results_ZIF268_MITOMI_df$Z.score_1_9, method = "spearman")

```



# compare to ProBound results
```{r}
# _NNNNN
All9mers_1 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers__NNNNN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_1) <- c("kmer", "rel_affinity1")

# N_NNNN
All9mers_2 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_N_NNNN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_2) <- c("kmer", "rel_affinity2")

# NN_NNN
All9mers_3 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NN_NNN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_3) <- c("kmer", "rel_affinity3")

# NNN_NN
All9mers_4 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NNN_NN_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_4) <- c("kmer", "rel_affinity4")

# NNNN_N
All9mers_5 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NNNN_N_ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_5) <- c("kmer", "rel_affinity5")

# NNNNN_
All9mers_6 <- read.table(paste(paste(Working_dir, "Input_Files/HT-SELEX_data_EGR1/ProBound", sep = "/"), "All_9mers_NNNNN__ProBound.txt", sep = "/"), header = FALSE)
colnames(All9mers_6) <- c("kmer", "rel_affinity6")

### Merge
All9mers <- All9mers_1
All9mers$rel_affinity2 <- All9mers_2$rel_affinity2
All9mers$rel_affinity3 <- All9mers_3$rel_affinity3
All9mers$rel_affinity4 <- All9mers_4$rel_affinity4
All9mers$rel_affinity5 <- All9mers_5$rel_affinity5
All9mers$rel_affinity6 <- All9mers_6$rel_affinity6
All9mers$trimmed_kmer <- substr(All9mers$kmer, 1, 9)
All9mers$rel_affinity <- apply(All9mers[, paste("rel_affinity", 1:6, sep = "")], 1, max)
ProBound_results <- All9mers[, c("trimmed_kmer", "rel_affinity")]

# Merge
TEMP1 <- merge(ZIF268_PBM_Escores_PADIT_subset, ProBound_results, 
               by.x = "X9.mer", by.y = "trimmed_kmer")
TEMP2 <- merge(TEMP1, ProBound_results, 
               by.x = "X9.mer.1", by.y = "trimmed_kmer")
ZIF268_PBM_Escores_PADIT_subset_ProBound <- TEMP2
ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity <- log2(apply(ZIF268_PBM_Escores_PADIT_subset_ProBound[, c("rel_affinity.x", "rel_affinity.y")], 1, max))

# Define threshold for log2_rel_affinity
log2_rel_affinity_threshold <- -5
ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig <- (ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity > log2_rel_affinity_threshold)*1

# Scatter plot 
ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour <- "black"
ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity == 1)] <- "red"
ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity == 1 & ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig == 1)] <- "cyan3"
plot(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour == "black")],
     ZIF268_PBM_Escores_PADIT_subset_ProBound$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour == "black")], 
     pch = 19, col = ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour == "black")], 
     ylim = c(-0.6, 3), xlim = c(-30, 0))
points(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour != "black")],
     ZIF268_PBM_Escores_PADIT_subset_ProBound$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour != "black")], 
     pch = 19, col = ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$Colour != "black")], 
     ylim = c(-0.6, 3))
abline(v = log2_rel_affinity_threshold, col = "blue")

# Boxplot
boxplot(ZIF268_PBM_Escores_PADIT_subset_ProBound$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig == 0 & ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity == 1)], 
        ZIF268_PBM_Escores_PADIT_subset_ProBound$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig == 1 & ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity == 1)], 
        notch = TRUE, outline = FALSE, col = c("red", "blue"))
wilcox.test(ZIF268_PBM_Escores_PADIT_subset_ProBound$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig == 0 & ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity == 1)], 
        ZIF268_PBM_Escores_PADIT_subset_ProBound$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig == 1 & ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity == 1)])
table(ZIF268_PBM_Escores_PADIT_subset_ProBound$log2_rel_affinity_sig, ZIF268_PBM_Escores_PADIT_subset_ProBound$MPRAactivity)

```



# compare to FIMO results
```{r}
# define colours
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour <- "black"
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity == 1)] <- "red"
ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$FIMO.p.value > -log10(0.0001))] <- "cyan3"

# plot
plot(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$E.score[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour == "black")],
     ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour == "black")], 
     cex = 0.5, pch = 19, col = ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour == "black")],  xlim = c(-0.5, 0.5), ylim = c(-0.5, 2.8))
points(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$E.score[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour != "black")],
     ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour != "black")], 
     cex = 0.5, pch = 19, col = ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour != "black")])
abline(h = 0, col = "red")

# boxplot
boxplot(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour == "red")],
        ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Mean_log2FC[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour == "cyan3" & ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity == 1)], 
        col = c("red", "cyan3"), notch = TRUE, outline = FALSE)
table(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour)
nrow(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO[which(ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$MPRAactivity == 1 & ZIF268_PBM_Escores_PADIT_HTSELEX_ProBound_FIMO$Colour == "cyan3"), ])

```
