==> 4.Compare_PADIT-seq_to_HOXD13_ChIP-seq_data.Rmd <==
---
title: "R Notebook"
output: html_notebook
---
# Set Global Variables here

```{r}
# Please change the working directory
Working_dir <- "~/Desktop/AVATAR/PADIT-seq-Manuscript/Code_Availability_Revision/"

# Please be sure to install these libraries
library(pheatmap); library(ggplot2); library(DESeq2); library(ggrepel); library("stringr"); library(viridis)

# FDR cutoff (do not change)
Cutoff_Value <- 0.05

```


# Important functions
```{r}
######################################################### Longest stretch of 1's
Longest_1s <- function(vec) {
  # Ensure that the vector contains only 0s and 1s
  if (!all(vec %in% c(0, 1))) {
    stop("The vector should contain only 0s and 1s")
  }

  # Use rle to compute the lengths of sequences of 1s and 0s
  runs <- rle(vec)

  # Extract the lengths of sequences of 1s only
  lengths_of_ones <- runs$lengths[runs$values == 1]

  # If there are no sequences of 1s, return a maximum length of 0 and start position as NA
  if (length(lengths_of_ones) == 0) {
    return(list(max_length = 0, start_position = NA))
  }

  # Identify the maximum length of the 1s sequences
  max_length <- max(lengths_of_ones)

  # Find the position (index) where the maximum stretch of 1s begins
  start_positions_of_ones <- cumsum(runs$lengths) - runs$lengths + 1
  start_position <- start_positions_of_ones[which(runs$lengths == max_length & runs$values == 1)[1]]

  # Return the maximum length and start position as a list
  return(list(max_length = max_length, start_position = start_position))
}

# Example usage:
vec <- c(1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1)
result <- Longest_1s(vec)
print(paste("Max length:", result$max_length, "| Start position:", result$start_position))

######################################################### Plot heatmap
plot_Heatmap <- function(Heatmap_DF_log2FC_rd_subset_subset, size)
{
  Heatmap_DF_log2FC_rd_subset_subset <- Heatmap_DF_log2FC_rd_subset_subset[which(apply(Heatmap_DF_log2FC_rd_subset_subset, 1, min) > -10), ]
  sub_mat <- Heatmap_DF_log2FC_rd_subset_subset[, 5:(ncol(Heatmap_DF_log2FC_rd_subset_subset)-4)]
  Heatmap_DF_log2FC_rd_subset_subset <- Heatmap_DF_log2FC_rd_subset_subset[order(apply(sub_mat, 1, sum), decreasing = TRUE), ]
  pheatmap(Heatmap_DF_log2FC_rd_subset_subset, 
         cluster_cols = FALSE, cluster_rows = FALSE, 
         show_rownames = FALSE, show_colnames =  FALSE, scale = "none", 
         color = inferno(100),
         main = paste(size, "bps", sep = " ")) 
}
```



# Read HOXD13 8-mer PADIT-seq log2foldchange values for the 3 registers
```{r}
################################################################# Read HOXD13 8-mer PADIT-seq log2foldchange values for the 3 registers
Register_1_8 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "HOXD13_all8mers_1_8_median.txt", sep = "/"), header = TRUE)
Register_2_9 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "HOXD13_all8mers_2_9_median.txt", sep = "/"), header = TRUE)
Register_3_10 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "HOXD13_all8mers_3_10_median.txt", sep = "/"), header = TRUE)

# Re-define MPRAactivity
Register_1_8$MPRAactivity <- (Register_1_8$log2FoldChange > 0 & Register_1_8$padj < Cutoff_Value)*1
Register_2_9$MPRAactivity <- (Register_2_9$log2FoldChange > 0 & Register_2_9$padj < Cutoff_Value)*1
Register_3_10$MPRAactivity <- (Register_3_10$log2FoldChange > 0 & Register_3_10$padj < Cutoff_Value)*1

# Number of red points?
table(Register_1_8$MPRAactivity)
table(Register_2_9$MPRAactivity)
table(Register_3_10$MPRAactivity)

# Format
colnames(Register_1_8) <- paste(colnames(Register_1_8), "1_8", sep = "_")
colnames(Register_2_9) <- paste(colnames(Register_2_9), "2_9", sep = "_")
colnames(Register_3_10) <- paste(colnames(Register_3_10), "3_10", sep = "_")
TEMP1 <- merge(Register_1_8, Register_2_9,
               by.x = "X8.mer_1_8", by.y = "X8.mer_2_9")
TEMP2 <- merge(TEMP1, Register_3_10,
               by.x = "X8.mer_1_8", by.y = "X8.mer_3_10")
HOXD13_PBM_Escores_PADIT <- TEMP2; rm(TEMP1); rm(TEMP2)
HOXD13_PBM_Escores_PADIT$Mean_log2FC <- apply(HOXD13_PBM_Escores_PADIT[,c("log2FoldChange_1_8", "log2FoldChange_2_9", "log2FoldChange_3_10")], 1, median)
HOXD13_PBM_Escores_PADIT$MPRAactivity <- apply(HOXD13_PBM_Escores_PADIT[,c("MPRAactivity_1_8", "MPRAactivity_2_9", "MPRAactivity_3_10")], 1, median)

#
df1 <- HOXD13_PBM_Escores_PADIT[,c("X8.mer_1_8", "Mean_log2FC", "MPRAactivity")]
colnames(df1) <- c("TFBS", "log2FoldChange", "MPRAactivity")
df2 <- HOXD13_PBM_Escores_PADIT[,c("X8.mer.1_1_8", "Mean_log2FC", "MPRAactivity")]
colnames(df2) <- c("TFBS", "log2FoldChange", "MPRAactivity")
dds_results_HOXD13_df <- rbind(df1, df2)
```


# Tile PADIT-seq HOXD13 8-mers on ChIP-seq peaks
```{r}
######################################################### Actual ChIP-seq peaks
# Read HOXD13 CHIP-seq DNA sequences
HOXD13_CHIP <- read.table(paste(paste(Working_dir, "Input_Files/ChIP-seq_data_HOXD13", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa.txt", sep = "/"), header = FALSE)
colnames(HOXD13_CHIP) <- c("V1")

# Obtain tiling
numPeaks <- nrow(HOXD13_CHIP)
Heatmap_DF_log2FC <- list()
Heatmap_DF_MPRAactivity <- list()
for(i in 1:numPeaks)
{
  CurrentPeak <- toupper(HOXD13_CHIP$V1[i])
  Sequence_10mers_DF <- data.frame()
  for(j in 1:(nchar(CurrentPeak)-7))
  {
    Sequence_10mers_DF[j, 1] <- substr(CurrentPeak, j, j+7)
  }
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df[which(dds_results_HOXD13_df$TFBS %in% Sequence_10mers_DF$V1),]
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df_subset[match(Sequence_10mers_DF$V1, dds_results_HOXD13_df_subset$TFBS),]
  Heatmap_DF_log2FC <- c(Heatmap_DF_log2FC, list(dds_results_HOXD13_df_subset$log2FoldChange))
  Heatmap_DF_MPRAactivity <- c(Heatmap_DF_MPRAactivity, list(dds_results_HOXD13_df_subset$MPRAactivity))
}

# Write
saveRDS(Heatmap_DF_log2FC, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_FDR_0.05.RData", sep = "/"))
saveRDS(Heatmap_DF_MPRAactivity, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_FDR_0.05.RData", sep = "/"))
rm(HOXD13_CHIP)

```



# Tile PADIT-seq k-mers on background sequences from Homer
```{r}
# Read EGR1 CHIP-seq DNA sequences
HOXD13_CHIP_OG <- read.table(paste(paste(Working_dir, "Input_Files/ChIP-seq_data_HOXD13", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa.txt", sep = "/"), header = FALSE)

# Read HOXD13 CHIP-seq DNA sequences
HOXD13_CHIP <- read.table(paste(paste(Working_dir, "Input_Files/ChIP-seq_data_HOXD13", sep = "/"), "seq.tsv", sep = "/"), header = FALSE)
HOXD13_CHIP <- HOXD13_CHIP[which(! HOXD13_CHIP$V2 %in% HOXD13_CHIP_OG$V1),]
HOXD13_CHIP <- HOXD13_CHIP[which(HOXD13_CHIP$V1 %in% c(paste(1:32999, "local+1", sep = "-")) ), ]
colnames(HOXD13_CHIP) <- c("V2", "V1")

# Obtain tiling
numPeaks <- nrow(HOXD13_CHIP)
Heatmap_DF_log2FC <- list()
Heatmap_DF_MPRAactivity <- list()
for(i in 1:numPeaks)
{
  CurrentPeak <- toupper(HOXD13_CHIP$V1[i])
  Sequence_10mers_DF <- data.frame()
  for(j in 1:(nchar(CurrentPeak)-7))
  {
    Sequence_10mers_DF[j, 1] <- substr(CurrentPeak, j, j+7)
  }
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df[which(dds_results_HOXD13_df$TFBS %in% Sequence_10mers_DF$V1),]
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df_subset[match(Sequence_10mers_DF$V1, dds_results_HOXD13_df_subset$TFBS),]
  Heatmap_DF_log2FC <- c(Heatmap_DF_log2FC, list(dds_results_HOXD13_df_subset$log2FoldChange))
  Heatmap_DF_MPRAactivity <- c(Heatmap_DF_MPRAactivity, list(dds_results_HOXD13_df_subset$MPRAactivity))
}

# Write
saveRDS(Heatmap_DF_log2FC, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_HomerBG_FDR_0.05.RData", sep = "/"))
saveRDS(Heatmap_DF_MPRAactivity, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_HomerBG_FDR_0.05.RData", sep = "/"))
rm(HOXD13_CHIP)

```



# Tile PADIT-seq HOXD13 8-mers on shuffled peaks (bedtools)
```{r}
# Read HOXD13 CHIP-seq DNA sequences
HOXD13_CHIP <- read.table(paste(paste(Working_dir, "Input_Files/ChIP-seq_data_HOXD13", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10_shuffled1.bed.fa.txt", sep = "/"), header = FALSE)
colnames(HOXD13_CHIP) <- c("V1")

# Obtain tiling
numPeaks <- nrow(HOXD13_CHIP)
Heatmap_DF_log2FC <- list()
Heatmap_DF_MPRAactivity <- list()
for(i in 1:numPeaks)
{
  CurrentPeak <- toupper(HOXD13_CHIP$V1[i])
  Sequence_10mers_DF <- data.frame()
  for(j in 1:(nchar(CurrentPeak)-7))
  {
    Sequence_10mers_DF[j, 1] <- substr(CurrentPeak, j, j+7)
  }
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df[which(dds_results_HOXD13_df$TFBS %in% Sequence_10mers_DF$V1),]
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df_subset[match(Sequence_10mers_DF$V1, dds_results_HOXD13_df_subset$TFBS),]
  Heatmap_DF_log2FC <- c(Heatmap_DF_log2FC, list(dds_results_HOXD13_df_subset$log2FoldChange))
  Heatmap_DF_MPRAactivity <- c(Heatmap_DF_MPRAactivity, list(dds_results_HOXD13_df_subset$MPRAactivity))
}

# Write
saveRDS(Heatmap_DF_log2FC, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_shuffled1_FDR_0.05.RData", sep = "/"))
saveRDS(Heatmap_DF_MPRAactivity, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_shuffled1_FDR_0.05.RData", sep = "/"))
rm(HOXD13_CHIP)
```


# Tile PADIT-seq HOXD13 8-mers on shuffled peaks (bedtools)
```{r}
# Read HOXD13 CHIP-seq DNA sequences
HOXD13_CHIP <- read.table(paste(paste(Working_dir, "Input_Files/ChIP-seq_data_HOXD13", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10_shuffled2.bed.fa.txt", sep = "/"), header = FALSE)
colnames(HOXD13_CHIP) <- c("V1")

# Obtain tiling
numPeaks <- nrow(HOXD13_CHIP)
Heatmap_DF_log2FC <- list()
Heatmap_DF_MPRAactivity <- list()
for(i in 1:numPeaks)
{
  CurrentPeak <- toupper(HOXD13_CHIP$V1[i])
  Sequence_10mers_DF <- data.frame()
  for(j in 1:(nchar(CurrentPeak)-7))
  {
    Sequence_10mers_DF[j, 1] <- substr(CurrentPeak, j, j+7)
  }
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df[which(dds_results_HOXD13_df$TFBS %in% Sequence_10mers_DF$V1),]
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df_subset[match(Sequence_10mers_DF$V1, dds_results_HOXD13_df_subset$TFBS),]
  Heatmap_DF_log2FC <- c(Heatmap_DF_log2FC, list(dds_results_HOXD13_df_subset$log2FoldChange))
  Heatmap_DF_MPRAactivity <- c(Heatmap_DF_MPRAactivity, list(dds_results_HOXD13_df_subset$MPRAactivity))
}

# Write
saveRDS(Heatmap_DF_log2FC, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_shuffled2_FDR_0.05.RData", sep = "/"))
saveRDS(Heatmap_DF_MPRAactivity, file=paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_shuffled2_FDR_0.05.RData", sep = "/"))
rm(HOXD13_CHIP)
```


# HOXD13 ChIP-seq peaks contain longer stretches of consecutively active HOXD13 8-mers
```{r}
# Read log2FC tiling
log2FC_peaks <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_FDR_0.05.RData", sep = "/"))
log2FC_HomerBG <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_HomerBG_FDR_0.05.RData", sep = "/"))
log2FC_Shuffled1 <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_shuffled1_FDR_0.05.RData", sep = "/"))
log2FC_Shuffled2 <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "log2FC_8bpstiling_HOXD13_peaks_shuffled2_FDR_0.05.RData", sep = "/"))

# Read MPRAactivity tiling
MPRAactivity_peaks <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_FDR_0.05.RData", sep = "/"))
MPRAactivity_HomerBG <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_HomerBG_FDR_0.05.RData", sep = "/"))
MPRAactivity_HomerBG <- rapply(MPRAactivity_HomerBG, f=function(x) ifelse(is.na(x),0,x), how="replace" )
MPRAactivity_Shuffled1 <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_shuffled1_FDR_0.05.RData", sep = "/"))
MPRAactivity_Shuffled2 <- readRDS(paste(paste(Working_dir, "Output_Files", sep = "/"), "MPRAactivity_8bpstiling_HOXD13_peaks_shuffled2_FDR_0.05.RData", sep = "/"))

######################################################### Longest 1's
# Longest 1's per peak
longest1s_perPeak <- sapply(MPRAactivity_peaks, function(col) Longest_1s(col)$max_length)
longest1s_perPeak_Shuffled_random <- sapply(MPRAactivity_HomerBG, function(col) Longest_1s(col)$max_length)
longest1s_perPeak_Shuffled_random1 <- sapply(MPRAactivity_Shuffled1, function(col) Longest_1s(col)$max_length)
longest1s_perPeak_Shuffled_random2 <- sapply(MPRAactivity_Shuffled2, function(col) Longest_1s(col)$max_length)

# Longest 1's per peak
# longest1s_perPeak <- (longest1s_perPeak+7)*((longest1s_perPeak != 0)*1)
# longest1s_perPeak_Shuffled_random <- (longest1s_perPeak_Shuffled_random+7)*((longest1s_perPeak_Shuffled_random != 0)*1)
# longest1s_perPeak_Shuffled_random2 <- (longest1s_perPeak_Shuffled_random2+7)*((longest1s_perPeak_Shuffled_random != 0)*1)

# Histogram (longest 1's of MPRAactivity) 
hist(longest1s_perPeak_Shuffled_random, breaks = 100, prob = FALSE, col = "black", ylim = c(0, 9000),
     xlim = c(0, 20), xlab = "Number of 8-mers with PADIT-seq activity", 
     main = "Background = Local")
hist(longest1s_perPeak, breaks = 100, prob = FALSE, col = "red",
     xlim = c(0, 20), add = TRUE)

# Histogram (longest 1's of MPRAactivity) 
hist(longest1s_perPeak_Shuffled_random1, breaks = 100, prob = FALSE, col = "black", ylim = c(0, 9000),
     xlim = c(0, 20), xlab = "Number of 8-mers with PADIT-seq activity", 
     main = "Background = Random length-matched genomic interval")
hist(longest1s_perPeak, breaks = 100, prob = FALSE, col = "red",
     xlim = c(0, 20), add = TRUE)

# 
table(longest1s_perPeak > 5)

# Wilcoxon test
wilcox.test(longest1s_perPeak, longest1s_perPeak_Shuffled_random)
wilcox.test(longest1s_perPeak, longest1s_perPeak_Shuffled_random1)
wilcox.test(longest1s_perPeak, longest1s_perPeak_Shuffled_random2)
wilcox.test(longest1s_perPeak_Shuffled_random, longest1s_perPeak_Shuffled_random1)
wilcox.test(longest1s_perPeak_Shuffled_random1, longest1s_perPeak_Shuffled_random2)

```


# Heatmap of consecutively active 8-mers
```{r}
# Read ChIP-seq peaks with DNA sequence
HOXD13_CHIP <- read.table(paste(paste(Working_dir, "Input_Files/ChIP-seq_data_HOXD13", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0.txt", sep = "/"))
HOXD13_CHIP <- HOXD13_CHIP[complete.cases(HOXD13_CHIP), ]
colnames(HOXD13_CHIP) <- c("chr", "start", "end", "DNA")

# Obtain tiling
numPeaks <- nrow(HOXD13_CHIP)
DF6 <- data.frame()
DF7 <- data.frame()
DF8 <- data.frame()
DF9 <- data.frame()
for(i in 1:numPeaks)
{
  CurrentPeak <- toupper(HOXD13_CHIP$DNA[i])
  Sequence_10mers_DF <- data.frame()
  for(j in 1:(nchar(CurrentPeak)-7))
  {
    Sequence_10mers_DF[j, 1] <- substr(CurrentPeak, j, j+7)
  }
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df[which(dds_results_HOXD13_df$TFBS %in% Sequence_10mers_DF$V1),]
  dds_results_HOXD13_df_subset <- dds_results_HOXD13_df_subset[match(Sequence_10mers_DF$V1, dds_results_HOXD13_df_subset$TFBS),]
  Longest_1s_results <- Longest_1s(dds_results_HOXD13_df_subset$MPRAactivity)
  if(Longest_1s_results$max_length == 6 & Longest_1s_results$start_position > 4 & (Longest_1s_results$start_position + 16) < nrow(dds_results_HOXD13_df_subset))
  {
    log2FC <- dds_results_HOXD13_df_subset$log2FoldChange[(Longest_1s_results$start_position-4):(Longest_1s_results$start_position+9)]
    DF6[i, 1] <- HOXD13_CHIP[i, 1]
    DF6[i, 2] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position - 3) - 1
    DF6[i, 3] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position + 12 + 3)
    DF6[i, 4] <- substr(HOXD13_CHIP$DNA[i], (Longest_1s_results$start_position-4), (Longest_1s_results$start_position+16))
    DF6[i, 5:18] <- as.numeric(log2FC)
  }
  if(Longest_1s_results$max_length == 7 & Longest_1s_results$start_position > 4 & (Longest_1s_results$start_position + 17) < nrow(dds_results_HOXD13_df_subset))
  {
    log2FC <- dds_results_HOXD13_df_subset$log2FoldChange[(Longest_1s_results$start_position-4):(Longest_1s_results$start_position+10)]
    DF7[i, 1] <- HOXD13_CHIP[i, 1]
    DF7[i, 2] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position - 3) - 1
    DF7[i, 3] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position + 13 + 3)
    DF7[i, 4] <- substr(HOXD13_CHIP$DNA[i], (Longest_1s_results$start_position-4), (Longest_1s_results$start_position+17))
    DF7[i, 5:19] <- as.numeric(log2FC)
  }
  if(Longest_1s_results$max_length == 8 & Longest_1s_results$start_position > 4 & (Longest_1s_results$start_position + 18) < nrow(dds_results_HOXD13_df_subset))
  {
    log2FC <- dds_results_HOXD13_df_subset$log2FoldChange[(Longest_1s_results$start_position-4):(Longest_1s_results$start_position+11)]
    DF8[i, 1] <- HOXD13_CHIP[i, 1]
    DF8[i, 2] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position - 3) - 1
    DF8[i, 3] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position + 14 + 3)
    DF8[i, 4] <- substr(HOXD13_CHIP$DNA[i], (Longest_1s_results$start_position-4), (Longest_1s_results$start_position+18))
    DF8[i, 5:20] <- as.numeric(log2FC)
  }
  if(Longest_1s_results$max_length == 9 & Longest_1s_results$start_position > 4 & (Longest_1s_results$start_position + 19) < nrow(dds_results_HOXD13_df_subset))
  {
    log2FC <- dds_results_HOXD13_df_subset$log2FoldChange[(Longest_1s_results$start_position-4):(Longest_1s_results$start_position+12)]
    DF9[i, 1] <- HOXD13_CHIP[i, 1]
    DF9[i, 2] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position - 3) - 1
    DF9[i, 3] <- HOXD13_CHIP[i, 2] + (Longest_1s_results$start_position + 15 + 3)
    DF9[i, 4] <- substr(HOXD13_CHIP$DNA[i], (Longest_1s_results$start_position-4), (Longest_1s_results$start_position+19))
    DF9[i, 5:21] <- as.numeric(log2FC)
  }
}
DF6 <- DF6[complete.cases(DF6), ]
DF7 <- DF7[complete.cases(DF7), ]
DF8 <- DF8[complete.cases(DF8), ]
DF9 <- DF9[complete.cases(DF9), ]

######################################################### Plot heatmaps
# Heatmap (log2FC) 6 consecutive
plot_Heatmap(data.matrix(DF6[, 5:ncol(DF6)]), "13")

# Heatmap (log2FC) 7 consecutive
plot_Heatmap(data.matrix(DF7[, 5:ncol(DF7)]), "14")

# Heatmap (log2FC) 8 consecutive
plot_Heatmap(data.matrix(DF8[, 5:ncol(DF8)]), "15")

# Heatmap (log2FC) 9 consecutive
plot_Heatmap(data.matrix(DF9[, 5:ncol(DF9)]), "16")

######################################################### Plot average across positions
# 6 consecutive
plot(1:ncol(DF6[, 5:ncol(DF6)]), 
     apply(DF6[, 5:ncol(DF6)], 2, mean), 
     pch = 19, type = "b", cex = 2, 
     col = c(rep("black", times = 4), rep("red", times = 6), rep("black", times = 4)))

# 7 consecutive
plot(1:ncol(DF7[, 5:ncol(DF7)]), 
     apply(DF7[, 5:ncol(DF7)], 2, mean), 
     pch = 19, type = "b", cex = 2, 
     col = c(rep("black", times = 4), rep("red", times = 7), rep("black", times = 4)))

# 8 consecutive
plot(1:ncol(DF8[, 5:ncol(DF8)]), 
     apply(DF8[, 5:ncol(DF8)], 2, mean), 
     pch = 19, type = "b", cex = 2, 
     col = c(rep("black", times = 4), rep("red", times = 9), rep("black", times = 4)))

# 9 consecutive
plot(1:ncol(DF9[, 5:ncol(DF9)]), 
     apply(DF9[, 5:ncol(DF9)], 2, mean), 
     pch = 19, type = "b", cex = 2, 
     col = c(rep("black", times = 4), rep("red", times = 9), rep("black", times = 4)))


######################################################### Write files to map PhastCons conservation scores
write.table(DF6, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_6consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), col.names = FALSE, row.names = FALSE, quote = FALSE, sep = "\t")
write.table(DF7, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_7consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), col.names = FALSE, row.names = FALSE, quote = FALSE, sep = "\t")
write.table(DF8, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_8consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), col.names = FALSE, row.names = FALSE, quote = FALSE, sep = "\t")
write.table(DF9, file = paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_9consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), col.names = FALSE, row.names = FALSE, quote = FALSE, sep = "\t")

```

# Generate logos
```{r}
library(seqLogo)

# Function to generate PWM logo for any specified k-mers
Generate_PWM_seqlogo <- function(kmers_vec, kmer_length = 8)
{
  # Generate data frame for seqlogo
  kmers_vec_df <- data.frame()
  for(i in 1:kmer_length)
  {
    letter_i <- substr(kmers_vec, i, i)
    count <- 0
    for(j in c("A", "C", "G", "T"))
    {
      count <- count + 1
      kmers_vec_df[i, count] <- sum(letter_i == j)
    }
  }
  colnames(kmers_vec_df) <- c("A", "C", "G", "T")
  
  #define function that divides the frequency by the row sum i.e. proportions
  proportion <- function(x){
     rs <- sum(x);
     return(x / rs);
  }
   
  #create position weight matrix
  HOXD13_seqlogo <- apply(kmers_vec_df, 1, proportion)
  HOXD13_seqlogo <- makePWM(HOXD13_seqlogo)
  seqLogo(HOXD13_seqlogo)
}

#
DF6 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_6consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), header = FALSE)
Generate_PWM_seqlogo(substr(DF6$V4, 2, 20), 19)

#
DF7 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_7consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), header = FALSE)
Generate_PWM_seqlogo(substr(DF7$V4, 2, 21), 20)

#
DF8 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_8consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), header = FALSE)
Generate_PWM_seqlogo(substr(DF8$V4, 2, 22), 21)

#
DF9 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "GSE81356.HOXD13.embryonic-forelimb-bud.mm10.bed.fa2.0_9consecutive8mers_and4bpsFlank_log2FC.bed", sep = "/"), header = FALSE)
Generate_PWM_seqlogo(substr(DF9$V4, 2, 23), 22)

```
