==> 6.Test_OverlappingBindingSites_Model_Pho4.nb.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Set Global Variables here

```{r}
# Please change the working directory
Working_dir <- "~/Desktop/AVATAR/PADIT-seq-Manuscript/Code_Availability_Revision/"

# Please be sure to install these libraries
library(pheatmap); library(ggplot2); library(DESeq2); library(ggrepel); library("stringr"); library(viridis)

# FDR cutoff (do not change)
Cutoff_Value <- 0.05
```



# Important functions
```{r}
######################################################### Longest stretch of 1's
Longest_1s <- function(vec) {
  # Ensure that the vector contains only 0s and 1s
  if (!all(vec %in% c(0, 1))) {
    stop("The vector should contain only 0s and 1s")
  }

  # Use rle to compute the lengths of sequences of 1s and 0s
  runs <- rle(vec)

  # Extract the lengths of sequences of 1s only
  lengths_of_ones <- runs$lengths[runs$values == 1]

  # If there are no sequences of 1s, return a maximum length of 0 and start position as NA
  if (length(lengths_of_ones) == 0) {
    return(list(max_length = 0, start_position = NA))
  }

  # Identify the maximum length of the 1s sequences
  max_length <- max(lengths_of_ones)

  # Find the position (index) where the maximum stretch of 1s begins
  start_positions_of_ones <- cumsum(runs$lengths) - runs$lengths + 1
  start_position <- start_positions_of_ones[which(runs$lengths == max_length & runs$values == 1)[1]]

  # Return the maximum length and start position as a list
  return(list(max_length = max_length, start_position = start_position))
}

# Example usage:
vec <- c(1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 0, 1)
result <- Longest_1s(vec)
print(paste("Max length:", result$max_length, "| Start position:", result$start_position))

```


# Read PHO4 8-mer PADIT-seq log2foldchange values for the 3 registers
```{r}
# Read PHO4 PBM data
PHO4_PBM_Escores_v1 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_PHO4", sep = "/"), "Pho4_Zhu_v1_8mers_11111111.txt", sep = "/"), header = TRUE)
PHO4_PBM_Escores_v2 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_PHO4", sep = "/"), "Pho4_Zhu_v2_8mers_11111111.txt", sep = "/"), header = TRUE)
PHO4_PBM_Escores_TEMP <- merge(PHO4_PBM_Escores_v1, PHO4_PBM_Escores_v2, by = "X8.mer")
PHO4_PBM_Escores_TEMP$E.score <- apply(PHO4_PBM_Escores_TEMP[, c("E.score.x", "E.score.y")], 1, median)
PHO4_PBM_Escores_TEMP$Z.score <- apply(PHO4_PBM_Escores_TEMP[, c("Z.score.x", "Z.score.y")], 1, median)
PHO4_PBM_Escores <- PHO4_PBM_Escores_TEMP[, c("X8.mer", "X8.mer.1.x", "E.score", "Z.score")]; rm(PHO4_PBM_Escores_TEMP)
colnames(PHO4_PBM_Escores) <- c("X8.mer", "X8.mer.1", "E.score", "Z.score")

# Read PHO4 8-mer PADIT-seq log2foldchange values for the 3 registers
Register_1_8 <- read.table(paste(paste(Working_dir, "Output_Files/Median_kmer_log2FC", sep = "/"), "PHO4_all8mers_1_8_median.txt", sep = "/"), header = TRUE)
Register_2_9 <- read.table(paste(paste(Working_dir, "Output_Files/Median_kmer_log2FC", sep = "/"), "PHO4_all8mers_2_9_median.txt", sep = "/"), header = TRUE)
Register_3_10 <- read.table(paste(paste(Working_dir, "Output_Files/Median_kmer_log2FC", sep = "/"), "PHO4_all8mers_3_10_median.txt", sep = "/"), header = TRUE)

# Re-define MPRAactivity
Register_1_8$MPRAactivity <- (Register_1_8$log2FoldChange > 0 & Register_1_8$padj < Cutoff_Value)*1
Register_2_9$MPRAactivity <- (Register_2_9$log2FoldChange > 0 & Register_2_9$padj < Cutoff_Value)*1
Register_3_10$MPRAactivity <- (Register_3_10$log2FoldChange > 0 & Register_3_10$padj < Cutoff_Value)*1

# Number of red points?
table(Register_1_8$MPRAactivity)
table(Register_2_9$MPRAactivity)
table(Register_3_10$MPRAactivity)

# Fraction of red points?
sum((Register_1_8$MPRAactivity == 1)*1)/nrow(Register_1_8)
sum((Register_2_9$MPRAactivity == 1)*1)/nrow(Register_2_9)
sum((Register_3_10$MPRAactivity == 1)*1)/nrow(Register_3_10)

################################################################# Average of all 3 registers
# Format
colnames(Register_1_8) <- paste(colnames(Register_1_8), "1_8", sep = "_")
colnames(Register_2_9) <- paste(colnames(Register_2_9), "2_9", sep = "_")
colnames(Register_3_10) <- paste(colnames(Register_3_10), "3_10", sep = "_")
TEMP1 <- merge(Register_1_8, Register_2_9, 
               by.x = "X8.mer_1_8", by.y = "X8.mer_2_9")
TEMP2 <- merge(TEMP1, Register_3_10, 
               by.x = "X8.mer_1_8", by.y = "X8.mer_3_10")
PHO4_PBM_Escores_PADIT <- TEMP2; rm(TEMP1); rm(TEMP2)

# Define Median values
PHO4_PBM_Escores_PADIT$Median_baseMean <- apply(PHO4_PBM_Escores_PADIT[,c("baseMean_1_8", "baseMean_2_9", "baseMean_3_10")], 1, median)
PHO4_PBM_Escores_PADIT$Mean_log2FC <- apply(PHO4_PBM_Escores_PADIT[,c("log2FoldChange_1_8", "log2FoldChange_2_9", "log2FoldChange_3_10")], 1, median)
PHO4_PBM_Escores_PADIT$padj <- apply(PHO4_PBM_Escores_PADIT[,c("padj_1_8", "padj_2_9", "padj_3_10")], 1, median)
PHO4_PBM_Escores_PADIT$MPRAactivity <- apply(PHO4_PBM_Escores_PADIT[,c("MPRAactivity_1_8", "MPRAactivity_2_9", "MPRAactivity_3_10")], 1, median)

# Account for the effect of flanking nucleotides
PHO4_PBM_Escores_PADIT$Median_baseMean <- ifelse(substr(PHO4_PBM_Escores_PADIT$X8.mer_1_8, 1, 4) == "ACGT" | substr(PHO4_PBM_Escores_PADIT$X8.mer.1_1_8, 1, 4) == "ACGT",
                                                 PHO4_PBM_Escores_PADIT$baseMean_2_9,
                                                 PHO4_PBM_Escores_PADIT$Median_baseMean)
PHO4_PBM_Escores_PADIT$Mean_log2FC <- ifelse(substr(PHO4_PBM_Escores_PADIT$X8.mer_1_8, 1, 4) == "ACGT" | substr(PHO4_PBM_Escores_PADIT$X8.mer.1_1_8, 1, 4) == "ACGT",
                                                 PHO4_PBM_Escores_PADIT$log2FoldChange_2_9,
                                                 PHO4_PBM_Escores_PADIT$Mean_log2FC)
PHO4_PBM_Escores_PADIT$padj <- ifelse(substr(PHO4_PBM_Escores_PADIT$X8.mer_1_8, 1, 4) == "ACGT" | substr(PHO4_PBM_Escores_PADIT$X8.mer.1_1_8, 1, 4) == "ACGT",
                                                 PHO4_PBM_Escores_PADIT$padj_2_9,
                                                 PHO4_PBM_Escores_PADIT$padj)
PHO4_PBM_Escores_PADIT$MPRAactivity <- ifelse(substr(PHO4_PBM_Escores_PADIT$X8.mer_1_8, 1, 4) == "ACGT" | substr(PHO4_PBM_Escores_PADIT$X8.mer.1_1_8, 1, 4) == "ACGT",
                                                 PHO4_PBM_Escores_PADIT$MPRAactivity_2_9,
                                                 PHO4_PBM_Escores_PADIT$MPRAactivity)

# Add PBM E-scores and Z-scores
TEMP <- merge(PHO4_PBM_Escores, PHO4_PBM_Escores_PADIT, 
              by.x = "X8.mer", by.y = "X8.mer_1_8")
PHO4_PBM_Escores_PADIT <- TEMP; rm(TEMP)

#
df1 <- PHO4_PBM_Escores_PADIT[,c("X8.mer", "Mean_log2FC", "MPRAactivity")]
colnames(df1) <- c("TFBS", "log2FoldChange", "MPRAactivity")
df2 <- PHO4_PBM_Escores_PADIT[,c("X8.mer.1", "Mean_log2FC", "MPRAactivity")]
colnames(df2) <- c("TFBS", "log2FoldChange", "MPRAactivity")
dds_results_PHO4_df <- rbind(df1, df2)

```


# Consecutive overlapping binding sites versus 60-mer PBM signal intensity
```{r}
# Read 60-mer PBM signal intensities
PHO4_PBM_All60mers <- read.delim(paste(paste(Working_dir, "Input_Files/PBM_data_PHO4", sep = "/"), "Pho4-9_rawdata.txt", sep = "/"), header = TRUE, sep = "\t")
PHO4_PBM_All60mers <- PHO4_PBM_All60mers[complete.cases(PHO4_PBM_All60mers),]
PHO4_PBM_All60mers <- PHO4_PBM_All60mers[which(substr(PHO4_PBM_All60mers$ID, 1, 3) %in% "dBr"), ]

######################################################### Tile probes
PHO4_PBM_All60mers$Longest_1s <- 0
PHO4_PBM_All60mers$SecondLongest_1s <- 0
PHO4_PBM_All60mers$ZeroesBetween <- -1
for(i in 1:nrow(PHO4_PBM_All60mers))
{
  Current_DNA_sequence <- PHO4_PBM_All60mers$Sequence[i]
  Sequence_8mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 7))
  {
    Sequence_8mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+7)
  }
  dds_results_PHO4_df_subset <- dds_results_PHO4_df[which(dds_results_PHO4_df$TFBS %in% Sequence_8mers_DF$V1),]
  dds_results_PHO4_df_subset <- dds_results_PHO4_df_subset[match(Sequence_8mers_DF$V1, dds_results_PHO4_df_subset$TFBS),]
  PHO4_PBM_All60mers$num_active[i] <- sum(dds_results_PHO4_df_subset$MPRAactivity)
  PHO4_PBM_All60mers$sum_log2FC[i] <- sum(2^dds_results_PHO4_df_subset$log2FoldChange)
  PHO4_PBM_All60mers$max_log2FC_active_only[i] <- max(2^dds_results_PHO4_df_subset$log2FoldChange)
  if(sum(dds_results_PHO4_df_subset$MPRAactivity) == 0)
  {
    PHO4_PBM_All60mers$sum_log2FC_active_only[i] <- 0
    PHO4_PBM_All60mers$max_log2FC_active_only[i] <- 0
    next
  }
  PHO4_PBM_All60mers$sum_log2FC_active_only[i] <- sum(2^dds_results_PHO4_df_subset$log2FoldChange[which(dds_results_PHO4_df_subset$MPRAactivity == 1)])
  
  # Longest and second longest 1's
  rle_1s <- rle(dds_results_PHO4_df_subset$MPRAactivity)
  longest_second_vec <- sort(rle_1s$lengths[which(rle_1s$values == 1)], decreasing = TRUE)
  if(length(longest_second_vec) > 0)
  {
    PHO4_PBM_All60mers$Longest_1s[i] <- longest_second_vec[1]
  }
  if(length(longest_second_vec) > 1)
  {
    PHO4_PBM_All60mers$SecondLongest_1s[i] <- longest_second_vec[2]
    
    # Distance between the highest and second highest
    longest_index <- which.max(rle_1s$lengths)
    second_index <- (sort(rle_1s$lengths, decreasing = TRUE))[2]
  
    if(longest_index < second_index) {
    # Second highest to the right
    zeros_between <- sum(rle_1s$lengths[(longest_index+1):(second_index-1)]) 
    } else {
    # Second highest to the left 
    zeros_between <- sum(rle_1s$lengths[(second_index+1):(longest_index-1)])
    }
    PHO4_PBM_All60mers$ZeroesBetween[i] <- zeros_between
  }
}

######################################################### Fit linear models
# subset to make sure there aren't any other active 9-mers
PHO4_PBM_All60mers_subset <- PHO4_PBM_All60mers[which(PHO4_PBM_All60mers$num_active == PHO4_PBM_All60mers$Longest_1s),]
# PHO4_PBM_All60mers_subset <- PHO4_PBM_All60mers_subset[which(PHO4_PBM_All60mers_subset$num_active > 0), ]

# run lm
summary(lm(log10(PHO4_PBM_All60mers_subset$Alexa488) ~ PHO4_PBM_All60mers_subset$sum_log2FC_active_only + PHO4_PBM_All60mers_subset$max_log2FC_active_only))
summary(lm(log10(PHO4_PBM_All60mers_subset$Alexa488) ~ PHO4_PBM_All60mers_subset$num_active + PHO4_PBM_All60mers_subset$max_log2FC_active_only))

# relative importance
library(relaimpo)
model <- lm(log10(Alexa488) ~ max_log2FC_active_only + num_active, data = PHO4_PBM_All60mers_subset)
model <- lm(log10(Alexa488) ~ max_log2FC_active_only + sum_log2FC_active_only, data = PHO4_PBM_All60mers_subset)
relative_importance <- calc.relimp(model, type = "lmg", rela = TRUE)
relative_importance

######################################################### Boxplot - 1
# subset to make sure there aren't any other active 8-mers
PHO4_PBM_All60mers_subset <- PHO4_PBM_All60mers[which(PHO4_PBM_All60mers$num_active == PHO4_PBM_All60mers$Longest_1s),]
PHO4_PBM_All60mers_subset$num_active[which(PHO4_PBM_All60mers_subset$num_active > 6)] <- 6
table(PHO4_PBM_All60mers_subset$num_active)
table(PHO4_PBM_All60mers$num_active)

# Boxplot
boxplot(log10(PHO4_PBM_All60mers_subset$Alexa488) ~ PHO4_PBM_All60mers_subset$num_active, 
        notch = TRUE, outline = FALSE)
pairwise.wilcox.test(PHO4_PBM_All60mers_subset$Alexa488,
                     PHO4_PBM_All60mers_subset$num_active)

######################################################### Boxplot - 2
# subset
PHO4_PBM_All60mers_subset <- PHO4_PBM_All60mers[which(PHO4_PBM_All60mers$Longest_1s + PHO4_PBM_All60mers$SecondLongest_1s == 5), ]
PHO4_PBM_All60mers_subset$interaction_term <- interaction(PHO4_PBM_All60mers_subset$Longest_1s, 
                                                            PHO4_PBM_All60mers_subset$SecondLongest_1s, 
                                                            drop = TRUE, lex.order = TRUE)

# Boxplot
boxplot(log10(PHO4_PBM_All60mers_subset$Alexa488) ~ droplevels(PHO4_PBM_All60mers_subset$interaction_term), 
        notch = TRUE, outline = FALSE)
table(PHO4_PBM_All60mers_subset$interaction_term)
pairwise.wilcox.test(PHO4_PBM_All60mers_subset$Alexa488,
                     PHO4_PBM_All60mers_subset$interaction_term)

```

# Consecutive overlapping 8-mer E-scores versus 10-mer PADIT-seq log2foldchange
```{r}
# Read PHO4 10-mer PADIT-seq data
dds_results_PHO4_df <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "dds_results_PHO4_df.txt", sep = "/"), header = TRUE)
dds_results_PHO4_df$MPRAactivity <- (dds_results_PHO4_df$log2FoldChange > 0 & dds_results_PHO4_df$padj < Cutoff_Value)*1

# Read PHO4 PBM data
PHO4_PBM_Escores_v1 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_PHO4", sep = "/"), "Pho4_Zhu_v1_8mers_11111111.txt", sep = "/"), header = TRUE)
PHO4_PBM_Escores_v2 <- read.table(paste(paste(Working_dir, "Input_Files/PBM_data_PHO4", sep = "/"), "Pho4_Zhu_v2_8mers_11111111.txt", sep = "/"), header = TRUE)
PHO4_PBM_Escores_TEMP <- merge(PHO4_PBM_Escores_v1, PHO4_PBM_Escores_v2, by = "X8.mer")
PHO4_PBM_Escores_TEMP$E.score <- apply(PHO4_PBM_Escores_TEMP[, c("E.score.x", "E.score.y")], 1, median)
PHO4_PBM_Escores_TEMP$Z.score <- apply(PHO4_PBM_Escores_TEMP[, c("Z.score.x", "Z.score.y")], 1, median)
PHO4_PBM_Escores <- PHO4_PBM_Escores_TEMP[, c("X8.mer", "X8.mer.1.x", "E.score", "Z.score")]; rm(PHO4_PBM_Escores_TEMP)
colnames(PHO4_PBM_Escores) <- c("X8.mer", "X8.mer.1", "E.score", "Z.score")

# Format PBM data
TEMP1 <- PHO4_PBM_Escores[,c("X8.mer", "E.score")]
TEMP2 <- PHO4_PBM_Escores[,c("X8.mer.1", "E.score")]
colnames(TEMP1) <- c("X8.mer", "E.score")
colnames(TEMP2) <- c("X8.mer", "E.score")
TEMP <- rbind(TEMP1, TEMP2)
PHO4_PBM_Escores <- TEMP

# 
dds_results_PHO4_df$mer8_1 <- substr(rownames(dds_results_PHO4_df), 1, 8)
dds_results_PHO4_df$mer8_2 <- substr(rownames(dds_results_PHO4_df), 2, 9)
dds_results_PHO4_df$mer8_3 <- substr(rownames(dds_results_PHO4_df), 3, 10)

# Merge
TEMP1 <- merge(dds_results_PHO4_df,
               PHO4_PBM_Escores, 
               by.x = "mer8_1", by.y = "X8.mer")
colnames(TEMP1)[ncol(TEMP1)] <- "E.score_1"
TEMP2 <- merge(TEMP1,
               PHO4_PBM_Escores, 
               by.x = "mer8_2", by.y = "X8.mer")
colnames(TEMP2)[ncol(TEMP2)] <- "E.score_2"
TEMP3 <- merge(TEMP2,
               PHO4_PBM_Escores, 
               by.x = "mer8_3", by.y = "X8.mer")
colnames(TEMP3)[ncol(TEMP3)] <- "E.score_3"
dds_results_PHO4_df_tiledEscores <- TEMP3

################################################################# Effect of flanking nucleotides (aggregate analysis)
correlation_test_results_df <- data.frame()
Unique_8mers <- unique(c(PHO4_PBM_Escores_PADIT$X8.mer[which(PHO4_PBM_Escores_PADIT$MPRAactivity == 1)], 
                         PHO4_PBM_Escores_PADIT$X8.mer.1[which(PHO4_PBM_Escores_PADIT$MPRAactivity == 1)]))

for(i in 1:length(Unique_8mers))
{
  dds_results_PHO4_df_tiledEscores_subset <- dds_results_PHO4_df_tiledEscores[which(dds_results_PHO4_df_tiledEscores$mer8_2 %in% Unique_8mers[i]),]
  if(nrow(dds_results_PHO4_df_tiledEscores_subset) < 8)
  {
    correlation_test_results_df[i, 1] <- 0
    correlation_test_results_df[i, 2] <- 1
    correlation_test_results_df[i, 3] <- 0
    correlation_test_results_df[i, 4] <- 0
    next
  }
  correlation_test_results <- cor.test(apply(dds_results_PHO4_df_tiledEscores_subset[, c("E.score_1", "E.score_3")], 1, mean), 
                                       dds_results_PHO4_df_tiledEscores_subset$log2FoldChange, 
                                       method = "pearson") 
  correlation_test_results_df[i, 1] <- correlation_test_results$estimate
  correlation_test_results_df[i, 2] <- correlation_test_results$p.value
  correlation_test_results_df[i, 3] <- nrow(dds_results_PHO4_df_tiledEscores_subset)
  correlation_test_results_df[i, 4] <- sum((dds_results_PHO4_df_tiledEscores_subset$MPRAactivity == 1)*1)
}
rownames(correlation_test_results_df) <- Unique_8mers
colnames(correlation_test_results_df) <- c("correlation_coeff", "p.value", "numObs", "numSig")

# Add PBM Escores to correlation_test_results_df
PHO4_PBM_Escores <- unique(PHO4_PBM_Escores)
TMP <- merge(correlation_test_results_df, 
              PHO4_PBM_Escores, 
              by.x = 0, 
              by.y = "X8.mer")
rownames(TMP) <- TMP$Row.names
correlation_test_results_df <- TMP[, -1]; rm(TMP)

# 
correlation_test_results_df <- correlation_test_results_df[which(correlation_test_results_df$numObs > 7),]
correlation_test_results_df <- correlation_test_results_df[which(correlation_test_results_df$numSig > 0),]
correlation_test_results_df$p.adj <- p.adjust(correlation_test_results_df$p.value, 
                                              method = "fdr")
correlation_test_results_df$p.adj.binary <- (correlation_test_results_df$p.adj<0.1)*1
table(correlation_test_results_df$p.adj.binary)

# Histogram - 1
hist(correlation_test_results_df$correlation_coeff, #[which(correlation_test_results_df$E.score > 0.4)], 
     breaks = 100, freq = FALSE,
     xlim = c(-1, 1), 
     col = "dark green",
     xlab = "Pearson R (mean E-scores of flanking nucleotides VS PHO4 PADIT-seq log2FC)", 
     main = "Effect of flanking nucleotides on PHO4 PADIT-seq log2FC")

################################################################# Effect of flanking nucleotides (case by case basis)
kmers_to_highlight <- c("ACACGTGA")
dds_results_PHO4_df_tiledEscores_subset2 <- unique(dds_results_PHO4_df_tiledEscores[which(dds_results_PHO4_df_tiledEscores$mer8_2 == kmers_to_highlight),])
dds_results_PHO4_df_tiledEscores_subset2$Colour <- "black"
dds_results_PHO4_df_tiledEscores_subset2$Colour[which(dds_results_PHO4_df_tiledEscores_subset2$MPRAactivity == 1)] <- "red"
plot(apply(dds_results_PHO4_df_tiledEscores_subset2[, c("E.score_1", "E.score_3")], 1, mean), 
     dds_results_PHO4_df_tiledEscores_subset2$log2FoldChange,
     cex = 3.0, pch = 19, 
     col = dds_results_PHO4_df_tiledEscores_subset2$Colour, 
     # xlim = c(0, 0.5), ylim = c(0, 3), 
     xlab = "Mean E-score", ylab = "PADIT-seq log2FC")
cor.test(apply(dds_results_PHO4_df_tiledEscores_subset2[, c("E.score_1", "E.score_3")], 1, mean), 
         dds_results_PHO4_df_tiledEscores_subset2$log2FoldChange)

################################################################# Linear model
# Subset
Unique_8mers <- unique(c(PHO4_PBM_Escores_PADIT$X8.mer_1_8[which(PHO4_PBM_Escores_PADIT$MPRAactivity == 1)], 
                         PHO4_PBM_Escores_PADIT$X8.mer.1_1_8[which(PHO4_PBM_Escores_PADIT$MPRAactivity == 1)]))
dds_results_PHO4_df_tiledEscores_subset3 <- dds_results_PHO4_df_tiledEscores[which(dds_results_PHO4_df_tiledEscores$mer8_1 %in% Unique_8mers | 
                                                                                       dds_results_PHO4_df_tiledEscores$mer8_2 %in% Unique_8mers | 
                                                                                       dds_results_PHO4_df_tiledEscores$mer8_3 %in% Unique_8mers),] 

# lm
summary(lm(dds_results_PHO4_df_tiledEscores_subset3$log2FoldChange ~ 
             dds_results_PHO4_df_tiledEscores_subset3$E.score_1 + 
             dds_results_PHO4_df_tiledEscores_subset3$E.score_2 + 
             dds_results_PHO4_df_tiledEscores_subset3$E.score_3))

summary(lm(dds_results_PHO4_df_tiledEscores_subset3$log2FoldChange ~ 
             dds_results_PHO4_df_tiledEscores_subset3$E.score_1))

summary(lm(dds_results_PHO4_df_tiledEscores_subset3$log2FoldChange ~ 
             dds_results_PHO4_df_tiledEscores_subset3$E.score_2))

summary(lm(dds_results_PHO4_df_tiledEscores_subset3$log2FoldChange ~ 
             dds_results_PHO4_df_tiledEscores_subset3$E.score_3))

```
