==> 11.Predict_AllelicSkew_PADIT-seq_MPRA.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Set Global Variables here

```{r}
# Please change the working directory
Working_dir <- "~/Desktop/AVATAR/PADIT-seq-Manuscript/Code_Availability_Revision/"

# Please be sure to install these libraries
library(pheatmap); library(ggplot2); library(DESeq2); library(ggrepel); library("stringr"); library(viridis)

# FDR cutoff (do not change)
Cutoff_Value <- 0.05

# Minimum Number of overlapping binding sites that need to differ between the two alleles?
MinNumOverlapBS <- 1
log2FC_cutoff <- 1

```


# Read EGR1 9-mer PADIT-seq log2foldchange values for the 2 registers
```{r}
# Read
Register_1_9 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "ZIF268_all9mers_1_9_median.txt", sep = "/"), header = TRUE)
Register_2_10 <- read.table(paste(paste(Working_dir, "Output_Files", sep = "/"), "ZIF268_all9mers_2_10_median.txt", sep = "/"), header = TRUE)

# Re-define MPRAactivity
Register_1_9$MPRAactivity <- (Register_1_9$log2FoldChange > 0 & Register_1_9$padj < Cutoff_Value)*1
Register_2_10$MPRAactivity <- (Register_2_10$log2FoldChange > 0 & Register_2_10$padj < Cutoff_Value)*1

# Format
colnames(Register_1_9) <- paste(colnames(Register_1_9), "1_9", sep = "_")
colnames(Register_2_10) <- paste(colnames(Register_2_10), "2_10", sep = "_")
TEMP1 <- merge(Register_1_9, Register_2_10,
               by.x = "X9.mer_1_9", by.y = "X9.mer_2_10")
ZIF268_PBM_Escores_PADIT <- TEMP1; rm(TEMP1)
ZIF268_PBM_Escores_PADIT$Mean_log2FC <- apply(ZIF268_PBM_Escores_PADIT[,c("log2FoldChange_1_9", "log2FoldChange_2_10")], 1, min)
ZIF268_PBM_Escores_PADIT$MPRAactivity <- apply(ZIF268_PBM_Escores_PADIT[,c("MPRAactivity_1_9", "MPRAactivity_2_10")], 1, min)

#
df1 <- ZIF268_PBM_Escores_PADIT[,c("X9.mer_1_9", "Mean_log2FC", "MPRAactivity")]
colnames(df1) <- c("TFBS", "log2FoldChange", "MPRAactivity")
df2 <- ZIF268_PBM_Escores_PADIT[,c("X9.mer.1_1_9", "Mean_log2FC", "MPRAactivity")]
colnames(df2) <- c("TFBS", "log2FoldChange", "MPRAactivity")
dds_results_EGR1_df <- rbind(df1, df2)
dds_results_EGR1_df$MPRAactivity[is.na(dds_results_EGR1_df$MPRAactivity)] <- 0

```


# Analyze MPRA data from Khetan et al. 
```{r}
# Read Allelic skew data
Supplementary_Table_3_All_SNPs_Tested_AllelicCalls <- read.table(paste(paste(Working_dir, "Input_Files/MPRA_data/Khetan", sep = "/"), "Khetan_2020_R2_Supplementary_Table_3_All_SNPs_Tested_AllelicCalls.txt", sep = "/"), header = TRUE)
Supplementary_Table_3_All_SNPs_Tested_AllelicCalls <- Supplementary_Table_3_All_SNPs_Tested_AllelicCalls[which(Supplementary_Table_3_All_SNPs_Tested_AllelicCalls$HigherAllele_DMSO != "-"), ] # Filter by allelic skew

# Read DNA sequences tested
data <- read.table(paste(paste(Working_dir, "Input_Files/MPRA_data/Khetan", sep = "/"), "T1D_khetan_All_20180111.asi.probes", sep = "/"), header = FALSE)
colnames(data) <- c("Allele_name", "SNP", "chr", "pos", "V5", "V6", "V7", "DNA_sequence")
data <- data[, c("SNP", "Allele_name", "chr", "pos", "DNA_sequence")]
data <- data[which(data$SNP %in% Supplementary_Table_3_All_SNPs_Tested_AllelicCalls$SNP),] # Filter 

######################################################### Tile over all SNPs with allelic skew
for(i in 1:nrow(data))
{
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  data$sum_PADITseq_active[i] <- sum(dds_results_EGR1_df_subset$MPRAactivity)
}

######################################################### Identify relevant SNPs
Unique_SNPs <- as.data.frame(unique(data$SNP[which(data$sum_PADITseq_active > 0)]))
colnames(Unique_SNPs) <- c("SNP")
for(i in 1:nrow(Unique_SNPs))
{
  Current_SNP_subset_df <- data[which(data$SNP %in% Unique_SNPs$SNP[i]),]
  if(length(unique(Current_SNP_subset_df$sum_PADITseq_active)) == 1)
  {
    Unique_SNPs$To_keep[i] <- 0
  }
  if(length(unique(Current_SNP_subset_df$sum_PADITseq_active)) >= MinNumOverlapBS)
  {
    Unique_SNPs$To_keep[i] <- 1
  }
}
# subset 
Unique_SNPs_of_interest <- Unique_SNPs[which(Unique_SNPs$To_keep == 1),]

######################################################### Format
Unique_SNPs_of_interest$To_Remove <- 0
for(i in 1:nrow(Unique_SNPs_of_interest))
{
  Current_SNP <- Unique_SNPs_of_interest$SNP[i]
  
  # Ref
  Current_DNA_sequence_ref <- data$DNA_sequence[which(data$Allele_name == paste(Current_SNP, "A", sep = "_"))]
  Sequence_9mers_DF_ref <- data.frame()
  
  # Alt
  if(nrow(data[which(data$SNP %in% Current_SNP),]) == 2)
  {
    # Alt
    Current_DNA_sequence_alt <- data$DNA_sequence[which(data$Allele_name == paste(Current_SNP, "B", sep = "_"))]
    Sequence_9mers_DF_alt <- data.frame()
  }
  
  if(nrow(data[which(data$SNP %in% Current_SNP),]) == 4)
  {
    # Alt
    Current_DNA_sequence_alt <- data$DNA_sequence[which(data$Allele_name == paste(Current_SNP, "B_alt-1", sep = "_"))]
    Sequence_9mers_DF_alt <- data.frame()
  }
  
  for(j in 1:(nchar(Current_DNA_sequence_ref) - 8))
  {
    Sequence_9mers_DF_ref[j, 1] <- substr(Current_DNA_sequence_ref, j, j+8)
    Sequence_9mers_DF_alt[j, 1] <- substr(Current_DNA_sequence_alt, j, j+8)
  }
  # ref
  dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF_ref$V1),]
  dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df_subset_ref[match(Sequence_9mers_DF_ref$V1, dds_results_EGR1_df_subset_ref$TFBS),]
  # alt
  dds_results_EGR1_df_subset_alt <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF_alt$V1),]
  dds_results_EGR1_df_subset_alt <- dds_results_EGR1_df_subset_alt[match(Sequence_9mers_DF_alt$V1, dds_results_EGR1_df_subset_alt$TFBS),]  
  
  # Merge
  colnames(dds_results_EGR1_df_subset_ref) <- paste(colnames(dds_results_EGR1_df_subset_ref), "ref", sep = "_")
  colnames(dds_results_EGR1_df_subset_alt) <- paste(colnames(dds_results_EGR1_df_subset_alt), "alt", sep = "_")
  dds_results_EGR1_df_subset <- cbind(dds_results_EGR1_df_subset_ref, dds_results_EGR1_df_subset_alt)
  
  # 
  if(abs(sum(dds_results_EGR1_df_subset$MPRAactivity_ref) - sum(dds_results_EGR1_df_subset$MPRAactivity_alt)) < MinNumOverlapBS)
  {
    Unique_SNPs_of_interest$To_Remove[i] <- 1
  }
  
  # Obtain difference
  dds_results_EGR1_df_subset$Diff_log2FC <- dds_results_EGR1_df_subset$log2FoldChange_alt - dds_results_EGR1_df_subset$log2FoldChange_ref
  
  # Store
  sum_ref <- sum(dds_results_EGR1_df_subset$log2FoldChange_ref[which(dds_results_EGR1_df_subset$MPRAactivity_ref == 1)])
  sum_alt <- sum(dds_results_EGR1_df_subset$log2FoldChange_alt[which(dds_results_EGR1_df_subset$MPRAactivity_alt == 1)])
  
  # 
  Unique_SNPs_of_interest$Diff_log2FC_alt_minus_ref[i] <- sum(dds_results_EGR1_df_subset$Diff_log2FC) # sum_alt - sum_ref
  Unique_SNPs_of_interest$Diff_active_alt_minus_ref[i] <- sum(dds_results_EGR1_df_subset$MPRAactivity_alt) - sum(dds_results_EGR1_df_subset$MPRAactivity_ref)
}

# Filter 
Unique_SNPs_of_interest <- Unique_SNPs_of_interest[which(Unique_SNPs_of_interest$To_Remove == 0),]
Unique_SNPs_of_interest <- Unique_SNPs_of_interest[which(abs(Unique_SNPs_of_interest$Diff_log2FC_alt_minus_ref) > log2FC_cutoff),]

# Merge with MPRA allelic calls
PADIT_MPRA_DF <- merge(Unique_SNPs_of_interest, Supplementary_Table_3_All_SNPs_Tested_AllelicCalls, by = "SNP")

# DMSO
PADIT_MPRA_DF$max_enhancer_activity <- apply(cbind(PADIT_MPRA_DF$REF_DMSO_MPRAactivity_log2FoldChange, PADIT_MPRA_DF$ALT_DMSO_MPRAactivity_log2FoldChange), 1, max)

# Concordant?
PADIT_MPRA_DF$Concordance <- (PADIT_MPRA_DF$Diff_log2FC_alt_minus_ref > 0 & PADIT_MPRA_DF$DMSO_AllelicSkew_LogFC > 0)*1 + (PADIT_MPRA_DF$Diff_log2FC_alt_minus_ref < 0 & PADIT_MPRA_DF$DMSO_AllelicSkew_LogFC < 0)*1
PADIT_MPRA_DF$Concordance[which(PADIT_MPRA_DF$Concordance == 0)] <- -1
table(PADIT_MPRA_DF$Concordance)

# plot
plot(PADIT_MPRA_DF$Diff_log2FC_alt_minus_ref, 
     PADIT_MPRA_DF$DMSO_AllelicSkew_LogFC, 
     pch = 19)
abline(h = 0, col = "red")
abline(v = 0, col = "red")

# Save for meta-analysis
DF1 <- PADIT_MPRA_DF[, c("SNP", "Diff_log2FC_alt_minus_ref", "DMSO_AllelicSkew_LogFC", "Diff_active_alt_minus_ref", "Concordance")]
DF1$Shape <- 15 # square
colnames(DF1) <- c("SNP", "Diff_log2FC_alt_minus_ref", "MPRA_logFC", "Diff_active_alt_minus_ref", "Concordance", "Shape")

######################################################### Plot example
data_subset <- data[which(data$SNP %in% "rs150937899"),]
for(i in 1:nrow(data_subset))
{
  Current_DNA_sequence <- data_subset$DNA_sequence[i]
  for(j in 1:(nchar(data_subset$DNA_sequence[i]) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df_subset_ref[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset_ref$TFBS),]
  
  # plot
  dds_results_EGR1_df_subset_ref$Colour <- "black"
  dds_results_EGR1_df_subset_ref$Colour[which(dds_results_EGR1_df_subset_ref$MPRAactivity == 1)] <- "red"
  plot(1:nrow(dds_results_EGR1_df_subset_ref), 
       dds_results_EGR1_df_subset_ref$log2FoldChange, 
       pch = 19, type = "b", ylim = c(-0.5, 1.6),
       col = dds_results_EGR1_df_subset_ref$Colour)
}

```




# Analyze MPRA data from McAfee et al. 
```{r}
# Read Allelic skew data
Supplementary_Table_3_All_SNPs_Tested_AllelicCalls <- read.table(paste(paste(Working_dir, "Input_Files/MPRA_data/McAfee", sep = "/"), "SNPs_with_AllelicSkew.txt", sep = "/"), header = TRUE)

# Read DNA sequences tested
data <- read.table(paste(paste(Working_dir, "Input_Files/MPRA_data/McAfee", sep = "/"), "SNPs_with_allelic_skew_20bps.bed.fa2.0.txt", sep = "/"), header = FALSE)
colnames(data) <- c("chr", "start", "end", "rsID", "DNA_sequence")
data$DNA_sequence <- toupper(data$DNA_sequence)

# Merge
TEMP <- merge(data, Supplementary_Table_3_All_SNPs_Tested_AllelicCalls, by = "rsID")
data <- TEMP; rm(TEMP)

################################################################# Add DNA sequence for the alt allele
for(i in 1:nrow(data))
{
  current_start <- data$start[i]
  current_pos <- data$BP[i]
  current_A1 <- data$A1[i]
  current_A2 <- data$A2[i]
  if(substr(data$DNA_sequence[i], 20, 20 + nchar(current_A1) - 1) == current_A1)
  {
    data$DNA_sequence_alt[i] <- data$DNA_sequence[i]
    substr(data$DNA_sequence_alt[i], 20, 20 + nchar(current_A1) - 1) <- current_A2
  }
}

######################################################### Tile - 1
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(data))
{
  # Ref
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- data$DNA_sequence_alt[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}

######################################################### Subset
vec <- abs(sapply(MPRA_activity_ref, sum) - sapply(MPRA_activity_alt, sum))
data <- data[which(vec >= MinNumOverlapBS),]

######################################################### Tile again
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(data))
{
  # Ref
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- data$DNA_sequence_alt[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}
data$Diff_log2FC_alt_minus_ref <- sapply(log2FC_alt, sum) - sapply(log2FC_ref, sum)
data$Diff_active_alt_minus_ref <- sapply(MPRA_activity_alt, sum) - sapply(MPRA_activity_ref, sum)

# filter 
data <- data[which(abs(data$Diff_log2FC_alt_minus_ref) > log2FC_cutoff),]
data <- data[which(abs(data$Diff_active_alt_minus_ref) >= MinNumOverlapBS),]

# Concordant?
data$Concordance <- (data$Diff_log2FC_alt_minus_ref > 0 & data$MPRA_logFC > 0)*1 + (data$Diff_log2FC_alt_minus_ref < 0 & data$MPRA_logFC < 0)*1
data$Concordance[which(data$Concordance == 0)] <- -1
table(data$Concordance, abs(data$Diff_active_alt_minus_ref))

# 
plot(data$Diff_log2FC_alt_minus_ref, 
     data$MPRA_logFC, pch = 19)
abline(h = 0)
abline(v = 0)

# Save for meta-analysis
DF3 <- data[, c("rsID", "Diff_log2FC_alt_minus_ref", "MPRA_logFC", "Diff_active_alt_minus_ref", "Concordance")]
DF3$Shape <- 18 # Rhombus
colnames(DF3) <- c("SNP", "Diff_log2FC_alt_minus_ref", "MPRA_logFC", "Diff_active_alt_minus_ref", "Concordance", "Shape")

```


# Analyze MPRA data from Tewhey et al. 
```{r}
# Read Allelic skew data
Supplementary_Table_3_All_SNPs_Tested_AllelicCalls <- read.table(paste(paste(Working_dir, "Input_Files/MPRA_data/Tewhey", sep = "/"), "SNPs_with_allelic_Skew.txt", sep = "/"), header = TRUE)
Supplementary_Table_3_All_SNPs_Tested_AllelicCalls <- Supplementary_Table_3_All_SNPs_Tested_AllelicCalls[, c("SNP", "Direction", "LogSkew.Comb", "C.Skew.logP", "C.Skew.fdr")]

# Read DNA sequences tested
data <- read.delim(paste(paste(Working_dir, "Input_Files/MPRA_data/Tewhey", sep = "/"), "SNPs_with_allelic_skew_20bps.bed.fa2.0.txt", sep = "/"), header = FALSE, sep = "\t")
colnames(data) <- c("chr", "start", "end", "SNP", "A1", "A2", "DNA_sequence")
data$DNA_sequence <- toupper(data$DNA_sequence)

# Merge
TEMP <- merge(data, Supplementary_Table_3_All_SNPs_Tested_AllelicCalls, by = "SNP")
data <- TEMP; rm(TEMP)

################################################################# Add DNA sequence for the alt allele
for(i in 1:nrow(data))
{
  current_start <- data$start[i]
  current_pos <- data$BP[i]
  current_A1 <- data$A1[i]
  current_A2 <- data$A2[i]
  if(substr(data$DNA_sequence[i], 20, 20 + nchar(current_A1) - 1) == current_A1)
  {
    data$DNA_sequence_alt[i] <- data$DNA_sequence[i]
    substr(data$DNA_sequence_alt[i], 20, 20 + nchar(current_A1) - 1) <- current_A2
  }
  if(substr(data$DNA_sequence[i], 20, 20 + nchar(current_A1) - 1) != current_A1)
  {
    data$DNA_sequence_alt[i] <- NA
    # data$DNA_sequence_alt[i] <- data$DNA_sequence[i]
    # substr(data$DNA_sequence_alt[i], 20, 20 + nchar(current_A1) - 1) <- current_A2
  }
}
data <- data[complete.cases(data),]

######################################################### Tile
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(data))
{
  # Ref
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- data$DNA_sequence_alt[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}

######################################################### Subset
vec <- abs(sapply(MPRA_activity_ref, sum) - sapply(MPRA_activity_alt, sum))
data <- data[which(vec >= MinNumOverlapBS),]

######################################################### Tile again
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(data))
{
  # Ref
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- data$DNA_sequence_alt[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}
data$Diff_log2FC_alt_minus_ref <- sapply(log2FC_alt, sum) - sapply(log2FC_ref, sum)
data$Diff_active_alt_minus_ref <- sapply(MPRA_activity_alt, sum) - sapply(MPRA_activity_ref, sum)

# filter 
data <- data[which(abs(data$Diff_log2FC_alt_minus_ref) > log2FC_cutoff),]
data <- data[which(abs(data$Diff_active_alt_minus_ref) >= MinNumOverlapBS),]

# Concordant?
data$Concordance <- (data$Diff_log2FC_alt_minus_ref > 0 & data$LogSkew.Comb > 0)*1 + (data$Diff_log2FC_alt_minus_ref < 0 & data$LogSkew.Comb < 0)*1
data$Concordance[which(data$Concordance == 0)] <- -1
table(data$Concordance, abs(data$Diff_active_alt_minus_ref))

# 
plot(data$Diff_log2FC_alt_minus_ref, 
     data$LogSkew.Comb, pch = 19)
abline(h = 0)
abline(v = 0)

# Save for meta-analysis
DF2 <- data[, c("SNP", "Diff_log2FC_alt_minus_ref", "LogSkew.Comb", "Diff_active_alt_minus_ref", "Concordance")]
DF2$Shape <- 17 # triangle
colnames(DF2) <- c("SNP", "Diff_log2FC_alt_minus_ref", "MPRA_logFC", "Diff_active_alt_minus_ref", "Concordance", "Shape")

```



# Analyze MPRA data from Abell et al. 
```{r}
# Read Allelic skew data
data <- read.table(paste(paste(Working_dir, "Input_Files/MPRA_data/Abell", sep = "/"), "SNPs_MPRA_data_DNA_sequences.txt", sep = "/"), header = TRUE)
data <- data[which(data$padj_allele < 0.1),]
data <- data[complete.cases(data),]

######################################################### SNP_id
data_bed <- data[, c("chrom", "pos", "ref", "alt", "strand")]
data_bed$pos_minus1 <- data_bed$pos - 1
data_bed <- data_bed[, c("chrom", "pos_minus1", "pos", "ref", "alt", "strand")]
data_bed$V4 <- paste(paste(paste(data_bed$chrom, data_bed$pos, sep = ":"), data_bed$ref, sep = ":"), data_bed$alt, sep = ":")
data_bed$V5 <- 0
data_bed <- data_bed[, c("chrom", "pos_minus1", "pos", "V4", "V5", "strand")]

# strand
data_bed$strand_to_use <- "+"
data_bed <- data_bed[, c("chrom", "pos_minus1", "pos", "V4", "V5", "strand_to_use")]

# Add 
data$SNP_id <- data_bed$V4

######################################################### Tile
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(data))
{
  # Ref
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- data$DNA_sequence_alt[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}

######################################################### Subset
vec <- abs(sapply(MPRA_activity_ref, sum) - sapply(MPRA_activity_alt, sum))
data <- data[which(vec >= MinNumOverlapBS),]

######################################################### Tile again
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(data))
{
  # Ref
  Current_DNA_sequence <- data$DNA_sequence[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- data$DNA_sequence_alt[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}
data$Diff_log2FC_alt_minus_ref <- sapply(log2FC_alt, sum) - sapply(log2FC_ref, sum)
data$Diff_active_alt_minus_ref <- sapply(MPRA_activity_alt, sum) - sapply(MPRA_activity_ref, sum)

# filter 
data <- data[which(abs(data$Diff_log2FC_alt_minus_ref) > log2FC_cutoff),]
data <- data[which(abs(data$Diff_active_alt_minus_ref) >= MinNumOverlapBS),]

# Concordant?
data$Concordance <- (data$Diff_log2FC_alt_minus_ref > 0 & data$log2FoldChange_allele > 0)*1 + (data$Diff_log2FC_alt_minus_ref < 0 & data$log2FoldChange_allele < 0)*1
data$Concordance[which(data$Concordance == 0)] <- -1
table(data$Concordance)

# 
plot(data$Diff_log2FC_alt_minus_ref, 
     data$log2FoldChange_allele, pch = 19)
abline(h = 0)
abline(v = 0)

# Save for meta-analysis
DF4 <- data[, c("SNP_id", "Diff_log2FC_alt_minus_ref", "log2FoldChange_allele", "Diff_active_alt_minus_ref", "Concordance")]
DF4$Shape <- 19 # circle
colnames(DF4) <- c("SNP", "Diff_log2FC_alt_minus_ref", "MPRA_logFC", "Diff_active_alt_minus_ref", "Concordance", "Shape")

```


# plot specific examples
```{r}
SNPs_to_plot <- c("chr12:69754012:C:T", "chr9:77705570:T:C", "chr18:59738283:C:T", 
                  "chr2:99758020:G:GCGCAGGCGCACGCCCA", "chr10:72237693:G:A", "chr5:61581431:G:A", "chr19:2108929:C:A", "chr5:56247528:G:T", "chr18:59738283:C:T", 
                  "chr9:116334458:C:T", "chr9:37592143:A:T", "chr20:61464184:C:T", "")

data_subset <- data[which(data$SNP_id %in% "chr22:43356130:A:C"),]

# Reference
Current_DNA_sequence <- data_subset$DNA_sequence[1]
for(j in 1:(nchar(Current_DNA_sequence) - 8))
{
  Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
}
dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df_subset_ref[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset_ref$TFBS),]
  
# plot
dds_results_EGR1_df_subset_ref$Colour <- "black"
dds_results_EGR1_df_subset_ref$Colour[which(dds_results_EGR1_df_subset_ref$MPRAactivity == 1)] <- "red"
plot(1:nrow(dds_results_EGR1_df_subset_ref), 
       dds_results_EGR1_df_subset_ref$log2FoldChange, 
       pch = 19, type = "b", ylim = c(-0.5, 3),
       col = dds_results_EGR1_df_subset_ref$Colour)

# Alt
Current_DNA_sequence <- data_subset$DNA_sequence_alt[1]
for(j in 1:(nchar(Current_DNA_sequence) - 8))
{
  Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
}
dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
dds_results_EGR1_df_subset_ref <- dds_results_EGR1_df_subset_ref[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset_ref$TFBS),]
  
# plot
dds_results_EGR1_df_subset_ref$Colour <- "black"
dds_results_EGR1_df_subset_ref$Colour[which(dds_results_EGR1_df_subset_ref$MPRAactivity == 1)] <- "red"
plot(1:nrow(dds_results_EGR1_df_subset_ref), 
       dds_results_EGR1_df_subset_ref$log2FoldChange, 
       pch = 19, type = "b", ylim = c(-0.5, 3),
       col = dds_results_EGR1_df_subset_ref$Colour)


```




# Combine and plot
```{r}
# 
Combined_DF <- rbind(DF1, DF2, DF3, DF4)

# Keep only 1 for duplicated SNPs
SNPs_twice <- (table(Combined_DF$SNP) == 2)*1
SNPs_twice <- SNPs_twice[which(SNPs_twice == 1)]
SNPs_twice_df <- Combined_DF[which(Combined_DF$SNP %in% names(SNPs_twice)),]
Combined_DF <- rbind(Combined_DF[which(! Combined_DF$SNP %in% names(SNPs_twice)),], 
                     SNPs_twice_df[c(1, 3),])

# Correct definitions
Combined_DF$Concordance[which(Combined_DF$Diff_active_alt_minus_ref > 0 & Combined_DF$Diff_log2FC_alt_minus_ref < 0)] <- 0
Combined_DF$Concordance[which(Combined_DF$Diff_active_alt_minus_ref < 0 & Combined_DF$Diff_log2FC_alt_minus_ref > 0)] <- 0
Combined_DF <- Combined_DF[which(Combined_DF$Concordance != 0),]

# How many SNPs predicted to disrupt EGR1 by PADIT-seq?
length(unique(Combined_DF$SNP))

# Concordant?
Combined_DF$Concordance <- (Combined_DF$Diff_log2FC_alt_minus_ref > 0 & Combined_DF$MPRA_logFC > 0)*1 + (Combined_DF$Diff_log2FC_alt_minus_ref < 0 & Combined_DF$MPRA_logFC < 0)*1
Combined_DF$Concordance[which(Combined_DF$Concordance == 0)] <- -1
table(Combined_DF$Concordance)

#################### plot
Combined_DF$Colour <- "black"
Combined_DF$Colour[which(Combined_DF$Concordance == 1)] <- "red"
Combined_DF$Colour[which(Combined_DF$SNP %in% c("1KGTest_var15850_77705570", "1KGTest_var26692_73274181", "1KGTest_var27457_59738283"))] <- "blue" 
plot(Combined_DF$Diff_log2FC_alt_minus_ref, 
     Combined_DF$MPRA_logFC, 
     pch = Combined_DF$Shape, 
     col = Combined_DF$Colour, cex = 1.5,
     xlab = "PADIT-seq Activity (Alt - Ref)", 
     ylab = "MPRA Allelic skew (Alt - Ref)")
abline(h = 0, col = "black")
abline(v = 0, col = "black")

# 
cor.test(Combined_DF$Diff_log2FC_alt_minus_ref, 
         Combined_DF$MPRA_logFC)

# 
cor.test(abs(Combined_DF$Diff_log2FC_alt_minus_ref), 
         abs(Combined_DF$MPRA_logFC))

# Binomial test for concordant versus dis-cordant SNPs
binom.test(table(Combined_DF$Concordance)[2], nrow(Combined_DF))

# 
barplot(table(abs(Combined_DF$Diff_active_alt_minus_ref)), ylim = c(0, 60))

```


# Add MotifBreakR results
```{r}
# Read
MotifBreakR_Abell <- read.delim(paste(paste(Working_dir, "Input_Files/MPRA_data/MotifBreakR", sep = "/"), "Abell_motifBreakR_results.txt", sep = "/"), header = TRUE, sep = "\t")
MotifBreakR_Tewhey_Khetan_schizo <- read.delim(paste(paste(Working_dir, "Input_Files/MPRA_data/MotifBreakR", sep = "/"), "Tewhey_Khetan_schizo_motifBreakR_results.txt", sep = "/"), header = TRUE, sep = "\t")
MotifBreakR_results <- rbind(MotifBreakR_Abell, MotifBreakR_Tewhey_Khetan_schizo)
MotifBreakR_results <- MotifBreakR_results[which(MotifBreakR_results$effect == "strong"),]

# How many SNPs predicted to disrupt EGR1 by MotifBreakR?
length(unique(MotifBreakR_results$SNP_id))

# How many SNPs predicted to disrupt EGR1 by MotifBreakR are also predicted by PADIT-seq to disrupt EGR1 binding?
sum((Combined_DF$SNP %in% MotifBreakR_results$SNP_id)*1)

# Format
Unique_SNP_ids_motifBreakR <- unique(MotifBreakR_results$SNP_id)
MotifBreakR_results_formatted <- data.frame()
for(i in 1:length(Unique_SNP_ids_motifBreakR))
{
  Current_SNP <- MotifBreakR_results[which(MotifBreakR_results$SNP_id %in% Unique_SNP_ids_motifBreakR[i]), ]
  if(nrow(Current_SNP) == 1)
  {
    MotifBreakR_results_formatted <- rbind(MotifBreakR_results_formatted, 
                                           Current_SNP)
    next
  }
  if(nrow(Current_SNP) > 1)
  {
    Current_SNP <- Current_SNP[which(abs(Current_SNP$alleleDiff) == max(abs(Current_SNP$alleleDiff))),]
    MotifBreakR_results_formatted <- rbind(MotifBreakR_results_formatted, 
                                           Current_SNP[1,])
    next
  }
}

# Merge
TEMP <- merge(Combined_DF, MotifBreakR_results_formatted, by.x = "SNP", by.y = "SNP_id")
PADIT_MPRA_MotifBreakr_DF <- TEMP[, which(! colnames(TEMP) %in% c("Refpvalue", "Altpvalue", "snpPos", "alleleRef", "alleleAlt"))]


#################### Venn Diagram
library(VennDiagram)
draw.pairwise.venn(area1 = nrow(Combined_DF), 
                   area2 = nrow(MotifBreakR_results_formatted),
                   cross.area = nrow(PADIT_MPRA_MotifBreakr_DF), 
                   fill = c("red", "green"))

#################### MotifBreakR performs poorly when the effect size is lower
boxplot(abs(Combined_DF$Diff_log2FC_alt_minus_ref[which(Combined_DF$SNP %in% PADIT_MPRA_MotifBreakr_DF$SNP)]), 
        abs(Combined_DF$Diff_log2FC_alt_minus_ref[which(! Combined_DF$SNP %in% PADIT_MPRA_MotifBreakr_DF$SNP)]), 
        notch = TRUE, outline = FALSE, 
        ylab = "abs(PADIT−seq Activity (Alt − Ref))", 
        names = c("Yes break", "No break"))

wilcox.test(abs(Combined_DF$Diff_log2FC_alt_minus_ref[which(Combined_DF$SNP %in% PADIT_MPRA_MotifBreakr_DF$SNP)]), 
            abs(Combined_DF$Diff_log2FC_alt_minus_ref[which(! Combined_DF$SNP %in% PADIT_MPRA_MotifBreakr_DF$SNP)]))

#################### There is no difference in MotifBreakR effect size for SNPs not detected by PADIT-seq to alter EGR1 binding
boxplot(abs(MotifBreakR_results_formatted$alleleDiff[which(! MotifBreakR_results_formatted$SNP_id %in% PADIT_MPRA_MotifBreakr_DF$SNP)]),
        abs(MotifBreakR_results_formatted$alleleDiff[which(MotifBreakR_results_formatted$SNP_id %in% PADIT_MPRA_MotifBreakr_DF$SNP)]), 
        notch = TRUE, outline = FALSE, 
        names = c("Absent in PADIT", "Present in PADIT"), ylab = "MotifBreakR effect size")

wilcox.test(abs(MotifBreakR_results_formatted$alleleDiff[which(MotifBreakR_results_formatted$SNP_id %in% PADIT_MPRA_MotifBreakr_DF$SNP)]), 
            abs(MotifBreakR_results_formatted$alleleDiff[which(! MotifBreakR_results_formatted$SNP_id %in% PADIT_MPRA_MotifBreakr_DF$SNP)]))

```


# Compare MotifBreakr and num overlapping binding sites altered
```{r}
# 
Combined_DF$MotifBreakR_prediction <- (Combined_DF$SNP %in% MotifBreakR_results_formatted$SNP_id)

# 
hist(abs(Combined_DF$Diff_active_alt_minus_ref[which(Combined_DF$MotifBreakR_prediction == "TRUE")]), 
     col = "green", ylim = c(0, 50), xlim = c(0, 7), breaks = 20)
hist(abs(Combined_DF$Diff_active_alt_minus_ref[which(Combined_DF$MotifBreakR_prediction == "FALSE")]), 
     col = "blue", ylim = c(0, 50), xlim = c(0, 7), breaks = 20)

# 
wilcox.test(abs(Combined_DF$Diff_active_alt_minus_ref) ~ Combined_DF$MotifBreakR_prediction)
table(abs(Combined_DF$Diff_active_alt_minus_ref) , Combined_DF$MotifBreakR_prediction)

```

# ADASTRA data & MPRA data
```{r}
# Read ADASTRA data
ADASTRA_data <- read.delim(paste(paste(Working_dir, "Input_Files/ADASTRA_data", sep = "/"), "AD_ASTRA_EGR1_HUMAN.tsv", sep = "/"), header = TRUE, sep = "\t")

# Merge
ADASTRA_data_DF <- merge(Combined_DF, ADASTRA_data, by.x = "SNP", by.y = "ID")

# plot
plot(ADASTRA_data_DF$Diff_log2FC_alt_minus_ref, 
     ADASTRA_data_DF$es_mean_alt - ADASTRA_data_DF$es_mean_ref, 
     pch = 19, cex = 2)
abline(h = 0, col = "black")
abline(v = 0, col = "black")

```


# Read ADASTRA data, obtain DNA sequences, and tile PADIT-seq data
```{r}
# Read ADASTRA data
ADASTRA_data <- read.delim(paste(paste(Working_dir, "Input_Files/ADASTRA_data", sep = "/"), "AD_ASTRA_EGR1_HUMAN.tsv", sep = "/"), header = TRUE, sep = "\t")

######################################################### R
library(GenomicRanges)
library(BSgenome.Hsapiens.UCSC.hg38)  # Make sure you have this installed
library(Biostrings)

# Function to get flanking sequences
get_flanking_sequences <- function(snp_data, flank_size = 20) {
  # Create GRanges object
  gr <- GRanges(
    seqnames = snp_data$X.chr,
    ranges = IRanges(
      start = snp_data$pos - flank_size,
      end = snp_data$pos + flank_size
    ),
    strand = "+",
    # Store additional info
    ref = snp_data$ref,
    alt = snp_data$alt,
    pos = snp_data$pos,
    rsid = snp_data$ID
  )
  
  # Get sequences
  seqs <- getSeq(BSgenome.Hsapiens.UCSC.hg38, gr)
  
  # Calculate position in extracted sequence where SNP occurs
  snp_pos <- flank_size + 1
  
  # Create reference and alternate sequences
  results <- data.frame(
    rsid = snp_data$ID,
    chr = snp_data$X.chr,
    pos = snp_data$pos,
    ref = snp_data$ref,
    alt = snp_data$alt,
    ref_seq = sapply(seq_along(seqs), function(i) {
      seq <- as.character(seqs[i])
      paste0(
        substr(seq, 1, snp_pos-1),
        snp_data$ref[i],
        substr(seq, snp_pos+1, nchar(seq))
      )
    }),
    alt_seq = sapply(seq_along(seqs), function(i) {
      seq <- as.character(seqs[i])
      paste0(
        substr(seq, 1, snp_pos-1),
        snp_data$alt[i],
        substr(seq, snp_pos+1, nchar(seq))
      )
    })
  )
  
  return(results)
}

# Use the function
flanking_sequences <- get_flanking_sequences(ADASTRA_data, flank_size = 20)

# 
colnames(flanking_sequences) <- c("rsid", "chr", "pos", "ref", "alt", "DNA_REF", "DNA_ALT")

TEMP <- merge(flanking_sequences, ADASTRA_data, by.x = "rsid", by.y = "ID")
EGR1_SNP_SELEX <- TEMP; rm(TEMP)

######################################################### 
#
log2FC_ref <- list()
MPRA_activity_ref <- list()
log2FC_alt <- list()
MPRA_activity_alt <- list()

# 
for(i in 1:nrow(EGR1_SNP_SELEX))
{
  # Ref
  Current_DNA_sequence <- EGR1_SNP_SELEX$DNA_REF[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_ref[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_ref[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
                    
  # Alt
  Current_DNA_sequence <- EGR1_SNP_SELEX$DNA_ALT[i]
  Sequence_9mers_DF <- data.frame()
  for(j in 1:(nchar(Current_DNA_sequence) - 8))
  {
    Sequence_9mers_DF[j, 1] <- substr(Current_DNA_sequence, j, j+8)
  }
  dds_results_EGR1_df_subset <- dds_results_EGR1_df[which(dds_results_EGR1_df$TFBS %in% Sequence_9mers_DF$V1),]
  dds_results_EGR1_df_subset <- dds_results_EGR1_df_subset[match(Sequence_9mers_DF$V1, dds_results_EGR1_df_subset$TFBS),] 
  log2FC_alt[[i]] <- dds_results_EGR1_df_subset$log2FoldChange
  MPRA_activity_alt[[i]] <- dds_results_EGR1_df_subset$MPRAactivity
}

# Define 
EGR1_SNP_SELEX$Diff_MPRA_Activity <- sapply(MPRA_activity_ref, sum) - sapply(MPRA_activity_alt, sum)
EGR1_SNP_SELEX$Diff_log2FC <- sapply(log2FC_ref, sum) - sapply(log2FC_alt, sum)

# PADIT-seq predictions
EGR1_SNP_SELEX$PADIT_prediction <- 0
EGR1_SNP_SELEX$PADIT_prediction[which(EGR1_SNP_SELEX$Diff_MPRA_Activity != 0)] <- 1
EGR1_SNP_SELEX$PADIT_prediction[which(abs(EGR1_SNP_SELEX$Diff_log2FC) < 1)] <- 0

# Correct definitions
EGR1_SNP_SELEX$Diff_log2FC[which(abs(EGR1_SNP_SELEX$Diff_log2FC) > 90)] <- 0
EGR1_SNP_SELEX$Diff_log2FC[which(EGR1_SNP_SELEX$Diff_log2FC > 0 & EGR1_SNP_SELEX$Diff_MPRA_Activity < 0)] <- 0
EGR1_SNP_SELEX$Diff_log2FC[which(EGR1_SNP_SELEX$Diff_log2FC < 0 & EGR1_SNP_SELEX$Diff_MPRA_Activity > 0)] <- 0
EGR1_SNP_SELEX$Diff_MPRA_Activity[which(EGR1_SNP_SELEX$Diff_log2FC == 0)] <- 0
EGR1_SNP_SELEX$Diff_MPRA_Activity[which(EGR1_SNP_SELEX$Diff_log2FC == 0)] <- 0
EGR1_SNP_SELEX$PADIT_prediction[which(EGR1_SNP_SELEX$Diff_log2FC == 0)] <- 0
EGR1_SNP_SELEX$PADIT_prediction[which(EGR1_SNP_SELEX$Diff_log2FC == 0)] <- 0

```


# Plot ADASTRA data and PADIT-seq data
```{r}
# subset
EGR1_FDR_0.05 <- read.delim(paste(paste(Working_dir, "Input_Files/ADASTRA_data", sep = "/"), "AD_ASTRA_search_results_EGR1_FDR_0.05.tsv", sep = "/"), header = TRUE, sep = "\t")
EGR1_SNP_SELEX_FDR_05 <- EGR1_SNP_SELEX[which(EGR1_SNP_SELEX$rsid %in% EGR1_FDR_0.05$rsID),]

# table
table(EGR1_SNP_SELEX_FDR_05$PADIT_prediction, 
      EGR1_SNP_SELEX_FDR_05$motif_conc)

# subset
EGR1_SNP_SELEX_FDR_05 <- EGR1_SNP_SELEX_FDR_05[which(EGR1_SNP_SELEX_FDR_05$PADIT_prediction == 0),]
EGR1_SNP_SELEX_FDR_05 <- EGR1_SNP_SELEX_FDR_05[complete.cases(EGR1_SNP_SELEX_FDR_05),]

# Assign discordant 
EGR1_SNP_SELEX_FDR_05$PADIT_Concordance <- "red"
EGR1_SNP_SELEX_FDR_05$PADIT_Concordance[which(EGR1_SNP_SELEX_FDR_05$Diff_log2FC > 0 & (EGR1_SNP_SELEX_FDR_05$es_mean_ref - EGR1_SNP_SELEX_FDR_05$es_mean_alt) < 0)] <- "black"
EGR1_SNP_SELEX_FDR_05$PADIT_Concordance[which(EGR1_SNP_SELEX_FDR_05$Diff_log2FC < 0 & (EGR1_SNP_SELEX_FDR_05$es_mean_ref - EGR1_SNP_SELEX_FDR_05$es_mean_alt) > 0)] <- "black"

# plot
EGR1_SNP_SELEX_FDR_05$Colour <- "black"
EGR1_SNP_SELEX_FDR_05$Colour[which(EGR1_SNP_SELEX_FDR_05$motif_conc %in% c("Discordant", "Weak Discordant"))] <- "red"
plot(EGR1_SNP_SELEX_FDR_05$Diff_log2FC, 
     EGR1_SNP_SELEX_FDR_05$es_mean_ref - EGR1_SNP_SELEX_FDR_05$es_mean_alt, 
     pch = 19, cex = 1.5, 
     xlab = "PADIT-seq (REF - ALT)", 
     ylab = "AD-ASTRA (REF - ALT)", 
     col = EGR1_SNP_SELEX_FDR_05$PADIT_Concordance)
abline(h = 0, col = "black")
abline(v = 0, col = "black")

# correlation
cor.test(EGR1_SNP_SELEX_FDR_05$Diff_log2FC, 
         EGR1_SNP_SELEX_FDR_05$es_mean_ref - EGR1_SNP_SELEX_FDR_05$es_mean_alt)

```



Run MotifBreakR
```{r}
library(motifbreakR); library(BSgenome); library(MotifDb); library(GenomicRanges)
library(BSgenome.Hsapiens.UCSC.hg19); library(SNPlocs.Hsapiens.dbSNP155.GRCh37)

#################### 
snps.mb <- snps.from.rsid(rsid = EGR1_FDR_0.05$rsID,
                          dbSNP = SNPlocs.Hsapiens.dbSNP155.GRCh37,
                          search.genome = BSgenome.Hsapiens.UCSC.hg19)

results <- motifbreakR(snpList = snps.mb, filterp = TRUE,
                       pwmList = MotifDb[grep('egr1', values (MotifDb)$geneSymbol, ignore.case=TRUE)],
                       threshold = 1e-4,
                       method = "ic",
                       bkg = c(A=0.25, C=0.25, G=0.25, T=0.25),
                       BPPARAM = BiocParallel::SerialParam())

results_df <- as.data.frame(results, row.names = NULL)

######################################################### Keep only 1 prediction
Unique_SNP_ids_motifBreakR <- unique(results_df$SNP_id)
results_df_formatted <- data.frame()
for(i in 1:length(Unique_SNP_ids_motifBreakR))
{
  Current_SNP <- results_df[which(results_df$SNP_id %in% Unique_SNP_ids_motifBreakR[i]), ]
  if(nrow(Current_SNP) == 1)
  {
    results_df_formatted <- rbind(results_df_formatted, 
                                           Current_SNP)
    next
  }
  if(nrow(Current_SNP) > 1)
  {
    Current_SNP <- Current_SNP[which(abs(Current_SNP$alleleDiff) == max(abs(Current_SNP$alleleDiff))),]
    results_df_formatted <- rbind(results_df_formatted, 
                                           Current_SNP[1,])
    next
  }
}

#################### 
# subset
EGR1_FDR_0.05 <- read.delim(paste(paste(Working_dir, "Input_Files/ADASTRA_data", sep = "/"), "AD_ASTRA_search_results_EGR1_FDR_0.05.tsv", sep = "/"), header = TRUE, sep = "\t")
EGR1_SNP_SELEX_FDR_05 <- EGR1_SNP_SELEX[which(EGR1_SNP_SELEX$rsid %in% EGR1_FDR_0.05$rsID),]
EGR1_SNP_SELEX_FDR_05 <- EGR1_SNP_SELEX_FDR_05[which(EGR1_SNP_SELEX_FDR_05$PADIT_prediction == 1),]
EGR1_SNP_SELEX_FDR_05_MotifBreakR <- merge(EGR1_SNP_SELEX_FDR_05, results_df_formatted, 
                                           by.x = "rsid", by.y = "SNP_id", 
                                           all.x = TRUE, all.y = TRUE)
EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize[is.na(EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize)] <- 0

# plot
plot(-1*EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize, 
     EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_ref - EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_alt, 
     pch = 19, cex = 1, 
     xlab = "PADIT-seq (REF - ALT)", 
     ylab = "AD-ASTRA (REF - ALT)")
abline(h = 0, col = "black")
abline(v = 0, col = "black")

# 
cor.test(-1*EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize, 
         EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_ref - EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_alt)

#################### 
# subset
EGR1_FDR_0.05 <- read.delim(paste(paste(Working_dir, "Input_Files/ADASTRA_data", sep = "/"), "AD_ASTRA_search_results_EGR1_FDR_0.05.tsv", sep = "/"), header = TRUE, sep = "\t")
EGR1_SNP_SELEX_FDR_05 <- EGR1_SNP_SELEX[which(EGR1_SNP_SELEX$rsid %in% EGR1_FDR_0.05$rsID),]
EGR1_SNP_SELEX_FDR_05_MotifBreakR <- merge(EGR1_SNP_SELEX_FDR_05, results_df_formatted, by.x = "rsid", by.y = "SNP_id")

# Assign discordant 
EGR1_SNP_SELEX_FDR_05_MotifBreakR$PADIT_Concordance <- "red"
EGR1_SNP_SELEX_FDR_05_MotifBreakR$PADIT_Concordance[which(EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize < 0 & (EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_ref - EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_alt) < 0)] <- "black"
EGR1_SNP_SELEX_FDR_05_MotifBreakR$PADIT_Concordance[which(EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize > 0 & (EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_ref - EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_alt) > 0)] <- "black"

# plot
plot(-1*EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize, 
     EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_ref - EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_alt, 
     pch = 19, cex = 1.5, 
     xlab = "MOTIFBREAKR (REF - ALT)", 
     ylab = "AD-ASTRA (REF - ALT)", 
     col = EGR1_SNP_SELEX_FDR_05_MotifBreakR$PADIT_Concordance)
abline(h = 0, col = "black")
abline(v = 0, col = "black")

# 
table(EGR1_SNP_SELEX_FDR_05_MotifBreakR$PADIT_Concordance)
cor.test(-1*EGR1_SNP_SELEX_FDR_05_MotifBreakR$alleleEffectSize, 
         EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_ref - EGR1_SNP_SELEX_FDR_05_MotifBreakR$es_mean_alt)

```
